### 2024-09-22 1559b75
script to interact with arcgis rest api feature layers
