+++ src/digitaltwin/arcgis_rest_api.py
@@ -1,0 +1,253 @@
+# -*- coding: utf-8 -*-
+"""
+This script provides functions to interact with ArcGIS REST API feature layers, generate query parameters, and
+retrieve geographic data for a specified area of interest.
+"""
+
+import asyncio
+import logging
+from typing import List, Dict, Union, NamedTuple
+
+import aiohttp
+import geopandas as gpd
+import pandas as pd
+import requests
+
+log = logging.getLogger(__name__)
+
+
+class RecordCounts(NamedTuple):
+    """
+    Represents the record counts of the feature layer.
+
+    Attributes
+    ----------
+    max_record_count : int
+        The maximum number of records that will be returned per query.
+    total_record_count : int
+        The total number of records available in the feature layer.
+    """
+
+    max_record_count: int
+    total_record_count: int
+
+
+def get_feature_layer_record_counts(url: str) -> RecordCounts:
+    """
+    Retrieve the maximum and total record counts from the feature layer.
+
+    Parameters
+    ----------
+    url : str
+        The URL of the feature layer.
+
+    Returns
+    -------
+    RecordCounts
+        A named tuple containing the maximum and total record counts of the feature layer.
+
+    Raises
+    ------
+    RuntimeError
+        If there is an issue with retrieving the record counts from the feature layer.
+    """
+    # Set up parameters for the initial request to get the maximum record count
+    params = {"f": "json"}
+    response = requests.get(url=url, params=params)
+    # Extract the maximum record count from the response
+    max_record_count = response.json()["maxRecordCount"]
+    # Set up parameters for the second request to get the total record count
+    params["where"] = "1=1"
+    params["returnCountOnly"] = True
+    response = requests.get(url=f"{url}/query", params=params)
+    try:
+        # Extract the total record count from the response
+        total_record_count = response.json()["count"]
+    except KeyError as e:
+        # Raise a RuntimeError to indicate the API failure
+        raise RuntimeError("Failed to get the feature layer record counts.") from e
+    # Returns the maximum and total record counts of the feature layer
+    return RecordCounts(max_record_count, total_record_count)
+
+
+def gen_query_param_list(
+        url: str,
+        area_of_interest: gpd.GeoDataFrame = None,
+        output_sr: int = None) -> List[Dict[str, Union[str, int]]]:
+    """
+    Generate a list of API query parameters used to retrieve ArcGIS REST API data.
+
+    Parameters
+    ----------
+    url : str
+        The URL of the feature layer.
+    area_of_interest : gpd.GeoDataFrame = None
+        A GeoDataFrame representing the area of interest for data retrieval. If not provided, all data will be fetched.
+    output_sr : int = None
+        The EPSG code of the spatial reference system in which the requested data should be returned if no area of
+        interest is provided.
+
+    Returns
+    -------
+    List[Dict[str, Union[str, int]]]
+        A list of API query parameters used to retrieve ArcGIS REST API data.
+    """
+    # If no area_of_interest is provided and output_sr is not specified, default output_sr to 2193
+    if area_of_interest is None and output_sr is None:
+        output_sr = 2193
+    # Raise an error if output_sr is provided when area_of_interest is already given
+    if area_of_interest is not None and output_sr is not None:
+        raise ValueError("`output_sr` should not be provided when `area_of_interest` is given.")
+    # Retrieves the maximum and total record counts from the feature layer
+    max_record_count, total_record_count = get_feature_layer_record_counts(url)
+    # Base query parameters used in each API call
+    query_params_base = {
+        "where": "1=1",
+        "outFields": "*",
+        "outSR": output_sr,
+        "f": "geojson",
[…]
