### 2024-09-24 2a3f8be
Fix missing docker build files for www service.
