### 2024-09-24 02b948d
Revert "Set BG_Flood defaults to values that are not as useful but can run on any system"
This reverts commit faf725a15651b08f425b97f2fb7972e48339a84a.
