+++ floodresilience/run_all.py
@@ -67,11 +67,11 @@ DEFAULT_MODULES_TO_PARAMETERS = {
         "log_level": LogLevel.INFO
     },
     bg_flood_model: {
-        "output_timestep": 1,
-        "end_time": 2,
+        "output_timestep": 100,
+        "end_time": 900,
         "resolution": None,
         "mask": 9999,
-        "gpu_device": -1,
+        "gpu_device": 0,
         "small_nc": 0,
         "log_level": LogLevel.INFO
     }
