### 2024-09-25 e6606d8
Update k and z. Change rainfall pH default
