+++ src/pollution_model/run_medusa_2.py
@@ -216,12 +216,27 @@ def total_metal_load_roof(surface_area: float,
         case SurfaceType.CONCRETE_ROOF:
             b = [2, -2.8, 0.5, 0.217, 3.57, -0.09, 7, -3.73]
             c = [50, 2600, 0.1, 0.01, 1, -3.1, -0.007, 0.056]
+            # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
+            k = 0.00933
+            # Duration of the event measured by hours - Roof
+            # observed from the intra-event concentration sampling
+            z = 0.75
         case SurfaceType.COPPER_ROOF:
             b = [100, -2.8, 1.372, 0.217, 3.57, -1, 275, -3.3]
             c = [-0.1, 2, 0.1, 0.01, 0.8, -1.3, -0.007, 0.056]
+            # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
+            k = 0.00933
+            # Duration of the event measured by hours - Roof
+            # observed from the intra-event concentration sampling
+            z = 0.75
         case SurfaceType.GALVANISED_ROOF:
             b = [2, -2.8, 0.5, 0.217, 3.57, -0.09, 7, -3.73]
             c = [910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.990]
+            # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
+            k = 0.00933
+            # Duration of the event measured by hours - Roof
+            # observed from the intra-event concentration sampling
+            z = 0.75
         case _:
             raise ValueError(invalid_surface_error)
 
@@ -234,12 +249,6 @@ def total_metal_load_roof(surface_area: float,
     initial_zinc_concentration = c[0] * rainfall_ph + c[1] * c[2] * antecedent_dry_days ** c[3] * (
         c[4] * average_rain_intensity ** c[5])
     second_stage_zinc = c[6] * rainfall_ph + c[7]
-
-    # Define Z as per experimental data
-    z = 0.75
-
-    # Define K, the wash off coefficient.
-    k = 1
 
     # Initialise total copper and zinc loads as a guaranteed common factor
     total_copper_load = initial_copper_concentration * surface_area * 1 / k
