### 2024-09-26 96c912f
Remove DATA_DIR_REC
