### 2024-09-27 96fe2f7
Quickly draft the task how to join two tables of points and polygons using building_id
