+++ src/pollution_model/run_medusa_2.py
@@ -27,6 +27,37 @@ from src.pollution_model.pollution_tables import MedusaScenarios
 
 log = logging.getLogger(__name__)
 
+def merge_data(engine: Engine) -> gpd.GeoDataFrame:
+    """
+    Read and merge building data under points and polygons. Then store them into database.
+
+    Parameters
+    ----------
+    engine : Engine
+        The engine used to connect to the database.
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        A GeoDataFrame containing all the information in point and polygon files.
+    """
+    # Read point data
+    point_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/points.geojson")
+    # Read polygon data
+    polygon_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/polygons.geojson")
+
+    # Merge point and polygon data using building_Id
+    merge_data = polygon_data.merge(point_data, left_on="building_Id", right_on="building_Id", how="left")
+    # Remove OBJECTID_x (as it is equal to building_Id), and change name of OBJECTID_y into OBJECTID_point
+    merge_data = merge_data.drop(columns=["OBJECTID_x"])
+    merge_data = merge_data.rename(columns={"OBJECTID_y": "OBJECTID_point"})
+
+    # Store the sea level rise data to the database table
+    log.info(f"Adding ROOF table data to the database.")
+    merge_data.to_postgis('ROOF table', engine, index=False, if_exists="replace")
+
+    return merge_data
+
 
 # Enum strings are assigned as they are described in the original paper
 class SurfaceType(StrEnum):
