### 2024-09-27 #7266 @kthyng
Ok I see my comments from last time were:
> all the editors who voted wondered if there could be a live demo available and why there wasn't one.

> Despite your list of docs, they still seem a bit light, but perhaps that will come up more during review.

and otherwise we were ready to move forward when you wanted to do some further development. I'll add this to our waitlist for an editor now.
