### 2024-09-29 #7266 @LukeParky
> all the editors who voted wondered if there could be a live demo available and why there wasn't one.

I agree with this, but unfortunately this has not been possible so far, though it may change in the future.

> Despite your list of docs, they still seem a bit light, but perhaps that will come up more during review.

We have extended the docs since this point according to informal feedback from users and developers, but perhaps more will come up during review.
