+++ src/environmental/water_quality/surface_water_quality.py
@@ -1,0 +1,305 @@
+# -*- coding: utf-8 -*-
+"""
+Fetch and clean surface water quality data from ECAN, store it in the database, and
+retrieve it for the requested area of interest.
+"""  # noqa: D400
+
+import asyncio
+import logging
+from datetime import datetime
+from http import HTTPStatus
+from io import StringIO
+from typing import List, Optional
+
+import aiohttp
+import geopandas as gpd
+import pandas as pd
+from sqlalchemy.engine import Engine
+from sqlalchemy.sql import text
+
+from src.digitaltwin.tables import check_table_exists
+from src.environmental.water_quality.surface_water_sites import get_surface_water_sites_from_db
+
+log = logging.getLogger(__name__)
+
+
+def clean_raw_surface_water_quality(surface_water_quality: pd.DataFrame) -> pd.DataFrame:
+    """
+    Clean and process the raw surface water quality data retrieved from ECAN.
+
+    Parameters
+    ----------
+    surface_water_quality : pd.DataFrame
+        A DataFrame containing the raw surface water quality data retrieved from ECAN.
+
+    Returns
+    -------
+    pd.DataFrame
+        A DataFrame containing the cleaned and processed surface water quality data.
+    """
+    # Convert all column names to lowercase and replace spaces with underscores to ensure consistency
+    surface_water_quality.columns = surface_water_quality.columns.str.lower().str.replace(' ', '_')
+    # Convert specified columns to string type and strip any leading and trailing whitespace
+    surface_water_quality['site_id'] = surface_water_quality['site_id'].astype(str).str.strip()
+    surface_water_quality['sampleid'] = surface_water_quality['sampleid'].astype(str).str.strip()
+    surface_water_quality['value'] = surface_water_quality['value'].astype(str).str.strip()
+    surface_water_quality['units'] = surface_water_quality['units'].astype(str).str.strip()
+    surface_water_quality['measurement'] = surface_water_quality['measurement'].astype(str).str.strip()
+    # Convert the 'collection_date' column to datetime format using the specified format
+    surface_water_quality["collection_date"] = pd.to_datetime(
+        surface_water_quality["collection_date"],
+        format="%d-%b %Y %H:%M%p"
+    )
+    # Remove unnecessary trailing zeros from the 'value' column
+    surface_water_quality['value'] = (
+        surface_water_quality['value']
+        .apply(lambda x: x.rstrip('0').rstrip('.') if '.' in x else x)
+    )
+    # Replace string representations of 'nan' and '*' with None to standardize missing values
+    surface_water_quality = surface_water_quality.replace({'nan': None, '*': None})
+    return surface_water_quality
+
+
+async def _fetch_surface_water_quality_for_site(sem: asyncio.Semaphore, site_id: str) -> Optional[pd.DataFrame]:
+    """
+    Fetch surface water quality data from ECAN for the specified site.
+
+    Parameters
+    ----------
+    sem : asyncio.Semaphore
+        A semaphore to limit the number of concurrent requests.
+    site_id : str
+        The unique site ID of the surface water site for which the data is to be fetched.
+
+    Returns
+    -------
+    Optional[pd.DataFrame]
+        A DataFrame containing surface water quality data from ECAN for the requested site,
+        or `None` if the data is unavailable or the site ID is invalid.
+
+    Raises
+    ------
+    aiohttp.ClientResponseError
+        If the request fails with a non-internal server-related HTTP error.
+    """
+    # Construct the URL for fetching surface water quality data using the provided site ID
+    url = f"https://www.ecan.govt.nz/data/water-quality-data/exportallsample/{site_id}"
+    try:
+        async with sem:
+            async with aiohttp.ClientSession() as session:
[…]
