### 2024-10-01 58c4b36
Adding in temporary designs to check specs.
Signed-off-by: Arpo <arpo@vanishingpoint.studio>
