### 2024-10-02 8a6247e
add main script for water quality data
