+++ src/environmental/water_quality/main_water_quality.py
@@ -1,0 +1,65 @@
+# -*- coding: utf-8 -*-
+"""
+Main water quality script used to fetch and store water quality data from ECAN in the database,
+and retrieve it for the requested area of interest.
+"""  # noqa: D400
+
+import logging
+
+import geopandas as gpd
+
+from src.digitaltwin import setup_environment
+from src.digitaltwin.utils import LogLevel, setup_logging, get_catchment_area
+from src.environmental.water_quality import surface_water_sites, surface_water_quality
+
+log = logging.getLogger(__name__)
+
+
+def main(
+        selected_polygon_gdf: gpd.GeoDataFrame,
+        log_level: LogLevel = LogLevel.DEBUG) -> gpd.GeoDataFrame:
+    """
+    Fetch and store water quality data from ECAN in the database, and retrieve the latest requested
+    surface water quality data from the database for the specified catchment area.
+
+    Parameters
+    ----------
+    selected_polygon_gdf : gpd.GeoDataFrame
+        A GeoDataFrame representing the selected polygon, i.e., the catchment area.
+    log_level : LogLevel = LogLevel.DEBUG
+        The log level to set for the root logger. Defaults to LogLevel.DEBUG.
+        The available logging levels and their corresponding numeric values are:
+        - LogLevel.CRITICAL (50)
+        - LogLevel.ERROR (40)
+        - LogLevel.WARNING (30)
+        - LogLevel.INFO (20)
+        - LogLevel.DEBUG (10)
+        - LogLevel.NOTSET (0)
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        A GeoDataFrame containing the latest water quality data for each site within the requested catchment area.
+    """
+    # Set up logging with the specified log level
+    setup_logging(log_level)
+    # Connect to the database
+    engine = setup_environment.get_database()
+    # Get catchment area
+    catchment_area = get_catchment_area(selected_polygon_gdf, to_crs=2193)
+
+    # Fetch surface water site data using the ArcGIS REST API and store it in the database
+    surface_water_sites.store_surface_water_sites_to_db(engine)
+    # Fetch surface water quality data for the requested catchment area and store it in the database
+    surface_water_quality.store_surface_water_quality_to_db(engine, catchment_area)
+    # Retrieve the latest requested surface water quality data from the database for the specified catchment area
+    latest_swq = surface_water_quality.get_req_surface_water_quality_from_db(engine, catchment_area, year=2021)
+    return latest_swq
+
+
+if __name__ == "__main__":
+    sample_polygon = gpd.GeoDataFrame.from_file("selected_polygon.geojson")
+    main(
+        selected_polygon_gdf=sample_polygon,
+        log_level=LogLevel.DEBUG
+    )
