### 2024-10-02 872011d
add water quality module to run_all.py
