### 2024-10-02 93ec9ce
update get_req_surface_water_quality_from_db() docstring
