### 2024-10-02 11e0222
Update `SurfaceType`. However, not have information for g and f used in `dissolved_metal_load` yet. Also, k and z were updated in other task but not updated here.
