### 2024-10-03 f17735d
Update g and f. Coefficients k and z were not updated here.
