### 2024-10-03 08d8736
Edit the wrong coefficients, mainly for NON_METAL.
