### 2024-10-04 7278aa8
Create roof surface data and write into database. Have not added area of interest yet. This might cause the error.
