+++ src/pollution_model/run_medusa_2.py
@@ -6,6 +6,9 @@ TZn (total zinc), and DZn (dissolved zinc).
 DOI of the model paper: https://doi.org/10.3390/w12040969
 """
 
+import os
+from src.digitaltwin import tables
+
 import logging
 import math
 from enum import StrEnum
@@ -28,6 +31,39 @@ from src.pollution_model.pollution_tables import MedusaScenarios
 log = logging.getLogger(__name__)
 
 
+# def merge_data(engine: Engine) -> gpd.GeoDataFrame:
+#     """
+#     Read and merge building data under points and polygons. Then store them into database.
+#
+#     Parameters
+#     ----------
+#     engine : Engine
+#         The engine used to connect to the database.
+#
+#     Returns
+#     -------
+#     gpd.GeoDataFrame
+#         A GeoDataFrame containing all the information in point and polygon files.
+#     """
+#     # Read point data
+#     point_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/points.geojson")
+#     # Read polygon data
+#     polygon_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/polygons.geojson")
+#
+#     # Merge point and polygon data using building_Id
+#     merging_data = polygon_data.merge(point_data, left_on="building_Id", right_on="building_Id", how="left")
+#     # Remove OBJECTID_x (as it is equal to building_Id), and change name of OBJECTID_y into OBJECTID_point
+#     merging_data = merging_data.drop(columns=["OBJECTID_x"])
+#     merging_data = merging_data.rename(columns={"OBJECTID_y": "OBJECTID_point"})
+#
+#     # Store the ROOF table data to the database table
+#     log.info("Adding ROOF table data to the database.")
+#     merging_data.to_postgis('ROOF table', engine, index=False, if_exists="replace")
+#
+#     return merging_data
+
+
+
 def merge_data(engine: Engine) -> gpd.GeoDataFrame:
     """
     Read and merge building data under points and polygons. Then store them into database.
@@ -42,22 +78,65 @@ def merge_data(engine: Engine) -> gpd.GeoDataFrame:
     gpd.GeoDataFrame
         A GeoDataFrame containing all the information in point and polygon files.
     """
-    # Read point data
-    point_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/points.geojson")
-    # Read polygon data
-    polygon_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/polygons.geojson")
+    # Change the proj path to avoid
+    # ERROR 1: PROJ: proj_create_from_database: Open of /home/martinnguyen204/miniconda3/envs/digitaltwin/share/proj failed
+    os.environ['PROJ_LIB'] = "/home/martinnguyen204/miniconda3/envs/digitaltwin/share/proj"
 
-    # Merge point and polygon data using building_Id
-    merging_data = polygon_data.merge(point_data, left_on="building_Id", right_on="building_Id", how="left")
-    # Remove OBJECTID_x (as it is equal to building_Id), and change name of OBJECTID_y into OBJECTID_point
-    merging_data = merging_data.drop(columns=["OBJECTID_x"])
-    merging_data = merging_data.rename(columns={"OBJECTID_y": "OBJECTID_point"})
+    # Read roof surface points from outside
+    roof_surface_points = gpd.read_file(
+        "/home/martinnguyen204/Digital-Twin/points.geojson"
+    )
+    # Check if the table already exist in the database
+    if tables.check_table_exists(engine, "roof_surface_points"):
+        log.info(f"roof_surface_points data already exists in the database.")
+    else:
[…]
