### 2024-10-07 dc6b826
Fix the bug if no surface water quality data is available from the database and add NoSurfaceWaterQualityException
