### 2024-10-07 3427bbf
return empty GeoDataFrame instead of None
