### 2024-10-07 9bea31b
remove get_req_surface_water_quality_from_db() function
