+++ src/environmental/water_quality/surface_water_quality.py
@@ -293,64 +293,3 @@ def store_surface_water_quality_to_db(engine: Engine, catchment_area: gpd.GeoDat
         except NoSurfaceWaterSitesException:
             # If no sites are found, log a message indicating that no water quality data is available
             log.info("No surface water quality data found for the requested catchment area.")
-
-
-def get_req_surface_water_quality_from_db(
-        engine: Engine,
-        catchment_area: gpd.GeoDataFrame,
-        year: Optional[int] = None) -> gpd.GeoDataFrame:
-    """
-    Retrieve the latest requested surface water quality data from the database for the specified catchment area.
-
-    Parameters
-    ----------
-    engine : Engine
-        The engine used to connect to the database.
-    catchment_area : gpd.GeoDataFrame
-        A GeoDataFrame representing the catchment area.
-    year : Optional[int] = None
-        The year from which to include the latest surface water quality data for sites within the specified
-        catchment area. If not provided or set to `None`, all latest data will be included.
-
-    Returns
-    -------
-    gpd.GeoDataFrame
-        A GeoDataFrame containing the latest requested surface water quality data from the database for the
-        specified catchment area, or an empty GeoDataFrame if no records are found.
-    """
-    log.info("Retrieving requested surface water quality data from the database.")
-
-    try:
-        # Retrieve surface water quality data from the database for the requested catchment area
-        swq = get_surface_water_quality_from_db(engine, catchment_area)
-
-        # Check if the retrieved data is empty
-        if swq.empty:
-            # Return None if no data is available
-            log.info(
-                "No surface water quality data is available for the specified catchment area.")
-            return gpd.GeoDataFrame()
-        else:
-            # Group the data by site_id to obtain the latest collection_date for each site
-            latest_dates = swq.groupby('site_id')['collection_date'].max().reset_index()
-            # Merge to obtain all records with the latest collection_date for each site
-            requested_records = latest_dates.merge(swq, on=['site_id', 'collection_date'], how='left')
-            # Check if a specific year is provided to filter the data
-            if year is not None:
-                # Filter for records where the year of collection_date is greater than or equal to the requested year
-                requested_records = requested_records[requested_records['collection_date'].dt.year >= year]
-            # Convert the result into a GeoDataFrame and reset the index
-            requested_records = gpd.GeoDataFrame(requested_records).reset_index(drop=True)
-            # Log a message if no records are found
-            if requested_records.empty:
-                log.info(
-                    "No requested surface water quality data is available for the specified catchment area.")
-                return gpd.GeoDataFrame()
-            # Log a success message and return the records if found
-            log.info("Retrieved the requested surface water quality data from the database.")
-            return requested_records
-
-    except NoSurfaceWaterQualityException as e:
-        # Log a message to indicate that surface water quality data is absent in the database
-        log.info(e)
-        return gpd.GeoDataFrame()
