### 2024-10-07 c4fcca2
remove the geometry column from the `surface_water_quality` table
