### 2024-10-08 7213cd4
Remove `# noqa: D400` from the scripts and set it as default by adding it to `.flake`
