+++ src/dynamic_boundary_conditions/__init__.py
@@ -2,4 +2,4 @@
 """
 Package to handle data and operations specific to dynamic boundary conditions,
 such as rainfall, tide and sea level rise, and river.
-"""  # noqa: D400
+"""
