### 2024-10-09 a8ad48f
update get_surface_water_quality_from_db() function
