### 2024-10-10 7768b9b
Edit building_Id because: - There are two cases of duplicates: everything is the same, and everything is the same except for surface_type - Remove all NANs in building_Id - Remove NANs in surface_type due to using the small version of data - Edit other mistakes (wrong abbr of ZnC, miswriting polygons and points, etc.)
