### 2024-10-10 577f30f
Revert "update docstring"
This reverts commit bf7865ce722778d135b9ecb26b1e62398ba9b857.
