### 2024-10-10 7836383
fix pylint duplicate-code
