+++ src/environmental/water_quality/surface_water_quality.py
@@ -301,7 +301,7 @@ def serve_surface_water_quality() -> None:
     """
     Serve the geospatial data for surface water quality visualisation and API use.
     Joins the surface_water_quality table to its corresponding geometry feature table.
-    """
+    """  # pylint: disable=duplicate-code
     # Create geoserver workspace for surface water quality data
     db_name = EnvVariable.POSTGRES_DB
     workspace_name = f"{db_name}-surface-water"
