### 2024-10-11 7341fe4
add endpoint for updating the `surface_water_sites` table when needed
