+++ pywps.cfg
@@ -1,0 +1,53 @@
+[metadata:main]
+identification_title=PyWPS Demo server
+identification_abstract=PyWPS testing and development server. Do NOT use this server in production environment. You shall setup PyWPS as WSGI application for production. Please refer documentation for further detials.
+identification_keywords=WPS,GRASS,PyWPS, Demo, Dev
+identification_keywords_type=theme
+identification_fees=None
+identification_accessconstraints=None
+provider_name=PyWPS Developement team
+provider_url=http://pywps.org/'
+contact_name=Your Name
+contact_position=Developer
+contact_address=My Street
+contact_city=My City
+contact_stateorprovince=None
+contact_postalcode=000 00
+contact_country=World, Internet
+contact_phone=+00 00 11 22 33
+contact_fax=+00 99 88 77 66
+contact_email=info@yourdomain.org
+contact_url=http://pywps.org
+contact_hours=8:00-20:00UTC
+contact_instructions=Knock on the door
+contact_role=pointOfContact
+
+[server]
+maxsingleinputsize=1mb
+maxrequestsize=3mb
+url=http://localhost:5000/wps
+outputurl=http://localhost:5000/outputs/
+outputpath=outputs
+workdir=workdir
+wd_inp_subdir=inputs
+wd_out_subdir=outputs
+maxprocesses=10
+parallelprocesses=2
+
+[processing]
+mode=docker
+port_min=5050
+port_max=5070
+docker_img=container
+dckr_inp_dir=/pywps-flask/data
+dckr_out_dir=/pywps-flask/outputs
+
+[logging]
+level=INFO
+file=logs/pywps.log
+database=sqlite:///logs/pywps-logs.sqlite3
+format=%(asctime)s] [%(levelname)s] file=%(pathname)s line=%(lineno)s module=%(module)s function=%(funcName)s %(message)s
+
+
+[grass]
+gisbase=/usr/local/grass-7.3.svn/
