### 2024-10-16 53e5c1d
Use AnyValue so terria js accepts literalinputs
