### 2024-10-16 398951f
Render an arbitrary terria catalog from a WPS service
