### 2024-10-16 7334863
Output static MEDUSA outputs
