### 2024-10-16 b19b8df
Clean up the interface
