### 2024-10-16 cc18a88
- Edit flake8 and pylint - Next step: improve the speed and edit area_of_interest
