+++ src/pollution_model/run_medusa_2.py
@@ -6,8 +6,6 @@ TZn (total zinc), and DZn (dissolved zinc).
 DOI of the model paper: https://doi.org/10.3390/w12040969
 """
 
-import os
-from src.digitaltwin import tables
 
 import logging
 import math
@@ -22,6 +20,7 @@ from sqlalchemy.sql import text
 from src import geoserver
 from src.config import EnvVariable
 from src.digitaltwin import setup_environment
+from src.digitaltwin import tables
 from src.digitaltwin.tables import create_table, check_table_exists
 from src.digitaltwin.tables import execute_query
 from src.digitaltwin.utils import get_catchment_area, LogLevel, setup_logging
@@ -29,39 +28,6 @@ from src.pollution_model.pollution_tables import Medusa2ModelOutputBuildings, Me
 from src.pollution_model.pollution_tables import MedusaScenarios
 
 log = logging.getLogger(__name__)
-
-
-# def merge_data(engine: Engine) -> gpd.GeoDataFrame:
-#     """
-#     Read and merge building data under points and polygons. Then store them into database.
-#
-#     Parameters
-#     ----------
-#     engine : Engine
-#         The engine used to connect to the database.
-#
-#     Returns
-#     -------
-#     gpd.GeoDataFrame
-#         A GeoDataFrame containing all the information in point and polygon files.
-#     """
-#     # Read point data
-#     point_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/points.geojson")
-#     # Read polygon data
-#     polygon_data = gpd.read_file("/home/martinnguyen204/Digital-Twin/polygons.geojson")
-#
-#     # Merge point and polygon data using building_Id
-#     merging_data = polygon_data.merge(point_data, left_on="building_Id", right_on="building_Id", how="left")
-#     # Remove OBJECTID_x (as it is equal to building_Id), and change name of OBJECTID_y into OBJECTID_point
-#     merging_data = merging_data.drop(columns=["OBJECTID_x"])
-#     merging_data = merging_data.rename(columns={"OBJECTID_y": "OBJECTID_point"})
-#
-#     # Store the ROOF table data to the database table
-#     log.info("Adding ROOF table data to the database.")
-#     merging_data.to_postgis('ROOF table', engine, index=False, if_exists="replace")
-#
-#     return merging_data
-
 
 
 def merge_data(engine: Engine) -> gpd.GeoDataFrame:
@@ -100,10 +66,10 @@ def merge_data(engine: Engine) -> gpd.GeoDataFrame:
     roof_surface_points = roof_surface_points.dropna(subset=['building_id', 'deeplearn_subclass'])
     # Check if the table already exist in the database
     if tables.check_table_exists(engine, "roof_surface_points"):
-        log.info(f"roof_surface_points data already exists in the database.")
+        log.info("roof_surface_points data already exists in the database.")
     else:
         # Store the building_point_data to the database table
-        log.info(f"Adding roof_surface_points table to the database.")
+        log.info("Adding roof_surface_points table to the database.")
         roof_surface_points.to_postgis("roof_surface_points", engine, index=False, if_exists="replace")
 
     # Read roof surface polygons from outside
@@ -117,10 +83,10 @@ def merge_data(engine: Engine) -> gpd.GeoDataFrame:
     })
     # Check if the table already exist in the database
     if tables.check_table_exists(engine, "roof_surface_polygons"):
-        log.info(f"roof_surface_polygons data already exists in the database.")
[…]
