### 2024-10-17 7ff7619
- Added area_of_interest in - Switch between reading table and check if table exists. It should be to check if it exists first if not then read. - Next step: Break the codes down into smaller functions - Next step: Not yet speed up the code
