+++ src/pollution_model/run_medusa_2.py
@@ -44,92 +44,89 @@ def merge_data(engine: Engine) -> gpd.GeoDataFrame:
     gpd.GeoDataFrame
         A GeoDataFrame containing all the information in point and polygon files.
     """
-    # Read roof surface points from outside
-    # This data has the deeplearn_matclass with roof types we need
-    roof_surface_points = gpd.read_file(
-        "/home/martinnguyen204/Digital-Twin/points_fullversion.shp"
-    )
-    # Rename
-    roof_surface_points = roof_surface_points.rename(columns={
-        'deeplearn_': 'deeplearn_matclass',
-        'deeplear_1': 'deeplearn_metalconf',
-        'dvr_matcla': 'dvr_matclass',
-        'ensemble_m': 'ensemble_matclass',
-        'ensemble_1': 'ensemble_metalconf',
-        'decay_inde': 'decay_index',
-        'deeplear_2': 'deeplearn_subclass',
-        'subclass_c': 'subclass_conf',
-        'galvanised': 'galvanised_conf',
-        'building_I': 'building_id'
-    })
-    # Remove rows of building_Id and deeplearn_subclass that are NANs
-    roof_surface_points = roof_surface_points.dropna(subset=['building_id', 'deeplearn_subclass'])
     # Check if the table already exist in the database
     if tables.check_table_exists(engine, "roof_surface_points"):
         log.info("roof_surface_points data already exists in the database.")
     else:
+        # Read roof surface points from outside
+        # This data has the deeplearn_matclass with roof types we need
+        roof_surface_points = gpd.read_file(
+            "/home/martinnguyen204/Digital-Twin/points_fullversion.shp"
+        )
+        # Rename
+        roof_surface_points = roof_surface_points.rename(columns={
+            'deeplearn_': 'deeplearn_matclass',
+            'deeplear_1': 'deeplearn_metalconf',
+            'dvr_matcla': 'dvr_matclass',
+            'ensemble_m': 'ensemble_matclass',
+            'ensemble_1': 'ensemble_metalconf',
+            'decay_inde': 'decay_index',
+            'deeplear_2': 'deeplearn_subclass',
+            'subclass_c': 'subclass_conf',
+            'galvanised': 'galvanised_conf',
+            'building_I': 'building_id'
+        })
+        # Remove rows of building_Id and deeplearn_subclass that are NANs
+        roof_surface_points = roof_surface_points.dropna(subset=['building_id', 'deeplearn_subclass'])
         # Store the building_point_data to the database table
         log.info("Adding roof_surface_points table to the database.")
         roof_surface_points.to_postgis("roof_surface_points", engine, index=False, if_exists="replace")
 
-    # Read roof surface polygons from outside
-    roof_surface_polygons = gpd.read_file(
-        "/home/martinnguyen204/Digital-Twin/polygons_fullversion.shp"
-    )
-    # Rename building_Id to building_id
-    roof_surface_polygons = roof_surface_polygons.rename(columns={
-        'building_I': 'building_id',
-        'Shape_Leng': 'Shape_Length'
-    })
     # Check if the table already exist in the database
     if tables.check_table_exists(engine, "roof_surface_polygons"):
         log.info("roof_surface_polygons data already exists in the database.")
     else:
+        # Read roof surface polygons from outside
+        roof_surface_polygons = gpd.read_file(
+            "/home/martinnguyen204/Digital-Twin/polygons_fullversion.shp"
+        )
+        # Rename building_Id to building_id
+        roof_surface_polygons = roof_surface_polygons.rename(columns={
+            'building_I': 'building_id',
+            'Shape_Leng': 'Shape_Length'
+        })
[…]
