### 2024-10-17 f4f4cce
Add first draft of Ōtākaro DT MEDUSA module paper
