+++ paper/otakaro_dt/otakaro_dt.md
@@ -1,0 +1,160 @@
+---
+title: 'Ōtākaro Digital Twin - A case study in a modular open-source Framework for Geospatial Digital Twins'
+tags:
+  - digital twin
+  - stormwater
+  - pollution
+  - spatial
+  - geospatial
+  
+authors:
+  - name: Casey Li
+    orcid: 0009-0009-1707-8289
+    equal-contrib: true 
+    affiliation: "1, 2"
+
+  - name: Luke Parkinson
+    orcid: 0000-0003-3130-3445
+    equal-contrib: true 
+    affiliation: 1
+    
+  - name: Xander Cai
+    orcid: 0009-0008-9995-9778
+    affiliation: 1
+  
+  - name: Martin Nguyen
+    orcid: 0000-0003-4469-0149
+    affiliation: 1
+
+  - name: Matthew Wilson
+    orcid: 0000-0001-9459-6981
+    affiliation: 1
+    
+  - name: Greg Preston
+    affiliation: 2
+  
+  - name: Sam
+    affiliation: 3
+  - name: [@@]todo others as needed
+    
+affiliations:
+[@@]todo consider affiliations appropriate for this particular paper
+  - name: Geospatial Research Institute | Toi Hangarau, New Zealand
+    index: 1
+  - name: Building Innovation Partnership, New Zealand
+    index: 2
+  - name: Auckland
+
+date: 17 October 2024
+bibliography: paper.bib
+---
+[@@ Matt notes -- Use this as a case study, and focus on the OARCDT. Also describe how our changes have impacted the core digital twin by creating a modular framework to allow for more domains. cut out core digital things as much as possible]
+# Summary
+Why the Ōtākaro is in need of a digital twin...
+
+How our implementation fills that need...
+
+What our digital twin is and what it enables..... (brief)
+
+How the Ōtākaro implementation is gives mana to the Ōtākaro...
+
+How this impacted the core of the digital twin project (give this project a name)
+
+Overview of the digital twin project evolving from FReDT, to an open-source framework.
+
+
+There is a lack of open-source solutions to geospatial digital twins [@@ provide reference]
+
+# Aim
+[@@ XANDER's notes - provide more context about exactly what we have built, so it does not come across as fluff that has little real world impact]
+[@@ Cut down on this aim]
+
+Establish an accessible and innovative platform, to collect and communicate knowledge about the Ōtākaro River, serving as a tool to give voice to the Ōtākaro and its importance to the community. 
+
+The platform aims to foster collaboration to support and encourage: 
+
+* Holistic decision making & governance. 
+* Environmental health and resilience outcomes 
+* Representation of the mauri of the Ōtākaro and; 
+* Amplify the significance of the Ōtākaro to the region and people by allowing users to model potential future narratives and tell stories of the Ōtākaro’s past & present. 
+
+The Ōtākaro Digital Twin purpose is guided by the Te Mana o te Wai framework and aims to be adaptable to other locations and environments.  
+
+
+
+
+# Background
+[Introduce our digital twin framework...]
+
+The Ōtākaro/Avon River Corridor (ŌARC) in Christchurch, New Zealand is undergoing significant development as part of the ŌARC Regeneration Plan (ref). It is an initiative by central government (?) and Christchurch City Council (CCC) to restore the health of the river, and bring people back as part of the recovery after the 2011-20XX Christchurch Earthquake Sequence. [@@ tie this back to the language found in the council documents]
+
+After the earthquakes, XX households were displaced...
+
+The OARC Regeneration will... (revitalise wetlands, recreation areas, water treatment, biodiversity etc.)
+
+To aid in decision making CCC is making use of (DigitalTwinFramework)
+
+
+
+
+# Statement of need
+#### Current situation
[…]
