### 2024-10-17 89dbf00
Add more structure to oarcdt paper
