+++ paper/oarcdt/paper_oarcdt.bib
@@ -1,0 +1,57 @@
+@misc{OSDT,
+    author = {Casey Li and Luke Parkinson and Pooja Khosla and Rose Pearson and Xander Cai and Martin Nguyen and Aporini Deer},
+    title = {Open Spatial Digital Twin},
+    year = {1337},
+    url = {https://github.com/GeospatialResearch/Digital-Twins},
+    urldate = {March 18 2024}
+}
+
+@misc{FREDT,
+    author = {Luke Parkinson},
+    title = {Flood Resilience Digital Twin},
+    year = {1337},
+    url = {https://github.com/GeospatialResearch/Digital-Twins},
+}
+
+@misc{FREDT-JOSS,
+    author = {Luke Parkinson},
+    title = {FReDT JOSS},
+    year = {1337},
+}
+
+@misc{medusa,
+    author = {Frances Charters},
+    year = {1337}
+}
+
+@misc{medusa-module,
+    author={Sam Thompson and Martin Nguyen and Luke Parkinson and Casey Li and Frances Charters},
+    title={Digital Twin Module MEDUSA 2.0},
+    year = {1337},
+}
+
+@misc{oarc-regen-plan,
+    author={CCC},
+    title={OARC Regen Plan},
+    year={1337}
+}
+
+@misc{todo,
+    author={X Todo},
+    year={1337},
+}
+
+
+
+
+
+
+
+@misc{imfe-2022,
+    author={Siddorn, John},
+    title={An Information Management Framework for Environmental Digital Twins (IMFe)},
+    year={1337},
+    url={https://zenodo.org/records/7004351},
+    comments={This is a nice broad overview of things that should be considered when building and environmental digital twin, but does not provide any specifics.}
+}
+
