+++ paper/compile_paper.sh
@@ -1,5 +1,5 @@
 # Get the root directory of this repository (the parent dir of this script)
-PAPER_DIR="$(dirname "$(readlink -fm "$0")")"
+PAPER_DIR="$(dirname "$(readlink -fm "$0")")/oarcdt"
 
 rm -f "$PAPER_DIR"/paper.pdf
 
