### 2024-10-18 1c35a9c
- Broke the codes down into smaller functions - Next step: Not yet speed up the code
