### 2024-10-19 #7266 @pritamd47
> 👋🏼 @pritamd47 and @arbennett, Would you like to review this submission to the Journal for Open Source Software? Our reviews are checklist-driven and openly conducted on GitHub over a timeline of 4–6 weeks. Because the process is much more iterative and interactive than a traditional paper review, we would ask you to start within the next 2 weeks. Here are reviewer guidelines for reference: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
>
> Thanks for your consideration.

Hi, thanks for the invitation to review. I am new to reviewing for JOSS. Should I go ahead and review the paper and submit it in this thread, or do I submit it elsewhere? The reviewer checklist mentions leaving a comment here on completion of the review. I just wasn't sure where to submit the review. Any past examples would also be helpful in understanding how to submit the review.
