### 2024-10-20 8af5b22
update EnvVariable class - remove unused env variables
