+++ src/config.py
@@ -94,8 +94,6 @@ class EnvVariable:  # pylint: disable=too-few-public-methods
     TEST_DATABASE_INTEGRATION = _get_bool_env_variable("TEST_DATABASE_INTEGRATION", default=True)
 
     DATA_DIR = pathlib.Path(_get_env_variable("DATA_DIR"))
-    DATA_DIR_REC = pathlib.Path(_get_env_variable("DATA_DIR_REC"))
-    DATA_DIR_MODEL_OUTPUT = pathlib.Path(_get_env_variable("DATA_DIR_MODEL_OUTPUT"))
     DATA_DIR_GEOSERVER = pathlib.Path(_get_env_variable("DATA_DIR_GEOSERVER"))
     FLOOD_MODEL_DIR = pathlib.Path(_get_env_variable("FLOOD_MODEL_DIR"))
 
