+++ src/flood_model/bg_flood_model.py
@@ -20,7 +20,6 @@ from sqlalchemy.engine import Engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import text
 
-from src import config
 from src.config import EnvVariable
 from src.digitaltwin import setup_environment
 from src.digitaltwin.s3_connection import S3Manager
@@ -50,7 +49,7 @@ def get_valid_bg_flood_dir() -> pathlib.Path:
         If the BG-Flood Model directory is not found or is not a valid directory.
     """
     # Get the BG-Flood Model directory from the environment variable
-    bg_flood_dir = config.EnvVariable.FLOOD_MODEL_DIR
+    bg_flood_dir = EnvVariable.FLOOD_MODEL_DIR
     # Check if the directory exists and is a valid directory
     if bg_flood_dir.exists() and bg_flood_dir.is_dir():
         return bg_flood_dir
@@ -68,7 +67,7 @@ def get_new_model_output_path() -> pathlib.Path:
         The path to the BG-Flood model output file.
     """
     # Get the BG-Flood model output directory from the environment variable
-    model_output_dir = config.EnvVariable.DATA_DIR / "model_output"
+    model_output_dir = EnvVariable.DATA_DIR / "model_output"
     # Create the BG-Flood model output directory if it does not already exist
     model_output_dir.mkdir(parents=True, exist_ok=True)
     # Get the current timestamp in "YYYY_MM_DD_HH_MM_SS" format
