+++ src/dynamic_boundary_conditions/river/main_river.py
@@ -9,7 +9,7 @@ from typing import Union, Optional
 
 import geopandas as gpd
 
-from src import config
+from src.config import EnvVariable
 from src.digitaltwin import setup_environment
 from src.digitaltwin.utils import LogLevel, setup_logging, get_catchment_area
 from src.dynamic_boundary_conditions.river import (
@@ -70,7 +70,7 @@ def main(
     # Get catchment area
     catchment_area = get_catchment_area(selected_polygon_gdf, to_crs=2193)
     # BG-Flood Model Directory
-    bg_flood_dir = config.EnvVariable.FLOOD_MODEL_DIR
+    bg_flood_dir = EnvVariable.FLOOD_MODEL_DIR
     # Remove any existing river model inputs in the BG-Flood directory
     river_model_input.remove_existing_river_inputs(bg_flood_dir)
 
