### 2024-10-21 17b6d69
- Add a line of code to run a small dataset - Full datasets under shapefile formats were sent through team - Broke the codes down into smaller functions - Next step: Not yet speed up the code
