### 2024-10-21 71997c2
update s3_connection.py
