+++ src/pollution_model/run_medusa_2.py
@@ -6,7 +6,6 @@ TZn (total zinc), and DZn (dissolved zinc).
 DOI of the model paper: https://doi.org/10.3390/w12040969
 """
 
-
 import logging
 import math
 from enum import StrEnum
@@ -20,9 +19,7 @@ from sqlalchemy.sql import text
 from src import geoserver
 from src.config import EnvVariable
 from src.digitaltwin import setup_environment
-from src.digitaltwin import tables
-from src.digitaltwin.tables import create_table, check_table_exists
-from src.digitaltwin.tables import execute_query
+from src.digitaltwin.tables import create_table, check_table_exists, execute_query
 from src.digitaltwin.utils import get_catchment_area, LogLevel, setup_logging
 from src.pollution_model.pollution_tables import Medusa2ModelOutputBuildings, Medusa2ModelOutputRoads
 from src.pollution_model.pollution_tables import MedusaScenarios
@@ -30,9 +27,9 @@ from src.pollution_model.pollution_tables import MedusaScenarios
 log = logging.getLogger(__name__)
 
 
-def read_roof_surface_polygons(engine: Engine) -> gpd.GeoDataFrame:
+def read_roof_surface_points(engine: Engine) -> None:
     """
-    Read building data under polygons. Then store them into database.
+    Read building data under points. Then store them into database.
 
     Parameters
     ----------
@@ -40,14 +37,14 @@ def read_roof_surface_polygons(engine: Engine) -> gpd.GeoDataFrame:
         The engine used to connect to the database.
     """
     # Check if the table already exist in the database
-    if tables.check_table_exists(engine, "roof_surface_points"):
+    if check_table_exists(engine, "roof_surface_points"):
         log.info("roof_surface_points data already exists in the database.")
     else:
         # Read roof surface points from outside
         # This data has the deeplearn_matclass with roof types we need
-        roof_surface_points = gpd.read_file(
-            "/home/martinnguyen204/Digital-Twin/points_fullversion.shp"
-        )
+        roof_points_file = pathlib.Path("./tmp/medusa_shapefiles/ccc/points_fullversion.shp")
+        log.info(f"Reading roof surface points file {roof_points_file}.")
+        roof_surface_points = gpd.read_file(roof_points_file)
         # Rename
         roof_surface_points = roof_surface_points.rename(columns={
             'deeplearn_': 'deeplearn_matclass',
@@ -68,7 +65,7 @@ def read_roof_surface_polygons(engine: Engine) -> gpd.GeoDataFrame:
         roof_surface_points.to_postgis("roof_surface_points", engine, index=False, if_exists="replace")
 
 
-def read_roof_surface_points(engine: Engine) -> gpd.GeoDataFrame:
+def read_roof_surface_polygons(engine: Engine) -> None:
     """
     Read building data under points. Then store them into database.
 
@@ -78,13 +75,13 @@ def read_roof_surface_points(engine: Engine) -> gpd.GeoDataFrame:
         The engine used to connect to the database.
     """
     # Check if the table already exist in the database
-    if tables.check_table_exists(engine, "roof_surface_polygons"):
+    if check_table_exists(engine, "roof_surface_polygons"):
         log.info("roof_surface_polygons data already exists in the database.")
     else:
         # Read roof surface polygons from outside
-        roof_surface_polygons = gpd.read_file(
-            "/home/martinnguyen204/Digital-Twin/polygons_fullversion.shp"
-        )
+        roof_polygon_file = pathlib.Path("./tmp/medusa_shapefiles/ccc/polygons_fullversion.shp")
+        log.info(f"Reading roof surface polygons file {roof_polygon_file}.")
[…]
