### 2024-10-22 e503ce2
Form MEDUSA processing into naive vectorised format to increase speed
