+++ src/pollution_model/run_medusa_2.py
@@ -8,11 +8,13 @@ DOI of the model paper: https://doi.org/10.3390/w12040969
 
 import logging
 import math
+import pathlib
 from enum import StrEnum
 from typing import NamedTuple, Dict, Union, Optional
 from xml.sax import saxutils
 
 import geopandas as gpd
+import pandas as pd
 from sqlalchemy.engine import Engine
 from sqlalchemy.sql import text
 
@@ -595,6 +597,48 @@ def get_road_information(engine: Engine, area_of_interest: gpd.GeoDataFrame) -> 
     return gpd.GeoDataFrame(roads_medusa_info)
 
 
+def run_medusa_model_for_single_surface(row: gpd.GeoSeries, rainfall_event: MedusaRainfallEvent) -> gpd.GeoSeries:
+    """
+    Run the pollution model for one surface geometry.
+    Calculate the TSS, total metal load, and dissolved metal load and add to the surface. This runs for one rain event.
+
+    Parameters
+    ----------
+    row: gpd.GeoSeries
+        A Polygon containing the surface's geometry with surface type info to run MEDUSA model on.
+    rainfall_event: MedusaRainfallEvent
+        Rainfall event parameters for MEDUSA model.
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        The MEDUSA results for the given surfaces and rainfall_event.
+    """
+    surface_type = row["surface_type"]
+    if surface_type == SurfaceType.ASPHALT_ROAD:
+        surface_area = row.geometry.length * 5
+    else:
+        surface_area = row.geometry.area
+
+    curr_tss = compute_tss_roof_road(surface_area, rainfall_event, surface_type)
+
+    curr_total_copper, curr_total_zinc = total_metal_load_surface(surface_area,
+                                                                  rainfall_event,
+                                                                  surface_type,
+                                                                  curr_tss)
+
+    curr_dissolved_copper, curr_dissolved_zinc = dissolved_metal_load(total_copper_load=curr_total_copper,
+                                                                      total_zinc_load=curr_total_zinc,
+                                                                      surface_type=surface_type)
+    updated_values = {"total_suspended_solids": curr_tss,
+                      "total_copper": curr_total_copper,
+                      "total_zinc": curr_total_zinc,
+                      "dissolved_copper": curr_dissolved_copper,
+                      "dissolved_zinc": curr_dissolved_zinc}
+    row.update(updated_values)
+    return row
+
+
 def run_medusa_model_for_surface_geometries(surfaces: gpd.GeoDataFrame,
                                             rainfall_event: MedusaRainfallEvent) -> gpd.GeoDataFrame:
     """
@@ -613,29 +657,11 @@ def run_medusa_model_for_surface_geometries(surfaces: gpd.GeoDataFrame,
     gpd.GeoDataFrame
         The MEDUSA results for the given surfaces and rainfall_event.
     """
-    for feature_id, row in surfaces.iterrows():
-        surface_type = row["surface_type"]
-        if surface_type == SurfaceType.ASPHALT_ROAD:
-            surface_area = row.geometry.length * 5
-        else:
-            surface_area = row.geometry.area
-
-        curr_tss = compute_tss_roof_road(surface_area, rainfall_event, surface_type)
-
-        curr_total_copper, curr_total_zinc = total_metal_load_surface(surface_area,
-                                                                      rainfall_event,
-                                                                      surface_type,
-                                                                      curr_tss)
-
-        curr_dissolved_copper, curr_dissolved_zinc = dissolved_metal_load(total_copper_load=curr_total_copper,
-                                                                          total_zinc_load=curr_total_zinc,
[…]
