### 2024-10-22 5386b8d
Add tqdm progress bar to long-running pandas apply
