### 2024-10-22 8ce4321
Simplify data handling, reducing number of db saves and making file handling easier
