+++ src/pollution_model/run_medusa_2.py
@@ -30,7 +30,7 @@ from src.pollution_model.pollution_tables import MedusaScenarios
 log = logging.getLogger(__name__)
 
 
-def read_roof_surface_points(engine: Engine) -> None:
+def save_roof_surface_type_points_to_db(engine: Engine) -> None:
     """
     Read building data under points. Then store them into database.
 
@@ -45,9 +45,8 @@ def read_roof_surface_points(engine: Engine) -> None:
     else:
         # Read roof surface points from outside
         # This data has the deeplearn_matclass with roof types we need
-        roof_points_file = pathlib.Path("./tmp/medusa_shapefiles/ccc/points_fullversion.shp")
-        log.info(f"Reading roof surface points file {roof_points_file}.")
-        roof_surface_points = gpd.read_file(roof_points_file)
+        log.info(f"Reading roof surface points file {EnvVariable.ROOF_SURFACE_TYPE_POINTS_PATH}.")
+        roof_surface_points = gpd.read_file(EnvVariable.ROOF_SURFACE_TYPE_POINTS_PATH)
         # Rename
         roof_surface_points = roof_surface_points.rename(columns={
             'deeplearn_': 'deeplearn_matclass',
@@ -68,7 +67,7 @@ def read_roof_surface_points(engine: Engine) -> None:
         roof_surface_points.to_postgis("roof_surface_points", engine, index=False, if_exists="replace")
 
 
-def read_roof_surface_polygons(engine: Engine) -> None:
+def save_roof_surface_polygons_to_db(engine: Engine) -> None:
     """
     Read building data under points. Then store them into database.
 
@@ -82,9 +81,8 @@ def read_roof_surface_polygons(engine: Engine) -> None:
         log.info("roof_surface_polygons data already exists in the database.")
     else:
         # Read roof surface polygons from outside
-        roof_polygon_file = pathlib.Path("./tmp/medusa_shapefiles/ccc/polygons_fullversion.shp")
-        log.info(f"Reading roof surface polygons file {roof_polygon_file}.")
-        roof_surface_polygons = gpd.read_file(roof_polygon_file)
+        log.info(f"Reading roof surface polygons file {EnvVariable.ROOF_SURFACE_POLYGON_PATH}.")
+        roof_surface_polygons = gpd.read_file(EnvVariable.ROOF_SURFACE_POLYGON_PATH)
         # Rename building_Id to building_id
         roof_surface_polygons = roof_surface_polygons.rename(columns={
             'building_I': 'building_id',
@@ -93,59 +91,6 @@ def read_roof_surface_polygons(engine: Engine) -> None:
         # Store the building_point_data to the database table
         log.info("Adding roof_surface_polygons table to the database.")
         roof_surface_polygons.to_postgis("roof_surface_polygons", engine, index=False, if_exists="replace")
-
-
-def merge_roof_surface(engine: Engine) -> None:
-    """
-    Merge surface type points with roof surface polygons. Then store them into database.
-
-    Parameters
-    ----------
-    engine : Engine
-        The engine used to connect to the database.
-    """
-    # Read building data under polygon format
-    read_roof_surface_polygons(engine)
-
-    # Read building data under point format
-    read_roof_surface_points(engine)
-
-    # Check if the table already exist in the database
-    if check_table_exists(engine, "roof_surface"):
-        log.info("roof_surface data already exists in the database.")
-    else:
-        # Merge building points and polygons using inner join,
-        # because there will be some building_id in polygon table
[…]
