### 2024-10-22 5ed1123
Move functions for buildings closer together
