### 2024-10-22 92079d3
Added check for existing scenario, to speed up processing in that case.
