### 2024-10-22 13b476e
Match coefficients for ASPHALT_RROAD TSS to MEDUSA spreadsheet
