### 2024-10-22 bf1ccba
update and move validate_aws_env_vars()
