### 2024-10-22 50b72f0
Apply suggestions from code review
