### 2024-10-22 fbac418
Remove duplicated building_ids again, since it impacts the amount of area calculated, and also can cause bugs.
