### 2024-10-23 67e6d5c
Add default ROOF_SURFACE_DATASET_PATH to not break tests
