### 2024-10-23 9ea9b76
Use absolute values of medusa outputs until #276 is resolved
