### 2024-10-23 803308f
Fix incorrect default value for rainfall_ph
