+++ src/pollution_model/medusa_process_service.py
@@ -1,20 +1,29 @@
+import io
 import json
 import math
+import time
+from threading import Thread
 
+import geopandas as gpd
 from pywps import Process, LiteralInput, LiteralOutput, UOM, ComplexInput, ComplexOutput, FORMATS, Format
 from pywps.inout.literaltypes import AllowedValue, AnyValue
 from pywps.validator.allowed_value import ALLOWEDVALUETYPE, RANGECLOSURETYPE
+
+from src.digitaltwin.utils import LogLevel
+from src.pollution_model import run_medusa_2
 
 
 class MedusaProcessService(Process):
     def __init__(self):
         inputs = [
             LiteralInput("antecedentDryDays", "Antecedent Dry Days", data_type='float', allowed_values=AnyValue()),
-            LiteralInput("averageRainIntensity", "Average Rain Intensity (mm/hour)", data_type='float', allowed_values=AnyValue()),
+            LiteralInput("averageRainIntensity", "Average Rain Intensity (mm/hour)", data_type='float',
+                         allowed_values=AnyValue()),
             LiteralInput("eventDuration", "Event Duration (hours)", data_type='float', allowed_values=AnyValue()),
         ]
         outputs = [
-            ComplexOutput("roofs", "Output", supported_formats=[Format("application/vnd.terriajs.catalog-member+json")]),
+            ComplexOutput("roofs", "Output",
+                          supported_formats=[Format("application/vnd.terriajs.catalog-member+json")]),
             ComplexOutput("roads", "Output", supported_formats=[Format("application/vnd.terriajs.catalog-member+json")])
         ]
         super(MedusaProcessService, self).__init__(
@@ -23,23 +32,45 @@ class MedusaProcessService(Process):
             title="Medusa",
             inputs=inputs,
             outputs=outputs,
+            status_supported=True,
             store_supported=True
         )
 
     def _handler(self, request, response):
+        antecedent_dry_days = request.inputs['antecedentDryDays'][0].data
+        average_rain_intensity = request.inputs['averageRainIntensity'][0].data
+        event_duration = request.inputs['eventDuration'][0].data
+        area_of_interest = gpd.GeoDataFrame.from_file("selected_polygon.geojson")
+
+        scenario_id = run_medusa_2.main(
+            area_of_interest,
+            LogLevel.DEBUG,
+            antecedent_dry_days,
+            average_rain_intensity,
+            event_duration
+        )
+
         response.outputs['roofs'].data = json.dumps({
-            "type": "geojson",
-            "url": "http://localhost:5000/outputs/MEDUSA_Roofs.geojson",
+            "type": "wfs",
             "name": "MEDUSA Roof Surfaces",
-            "id": "some unique I111D"
-        })
-        response.outputs['roads'].data = json.dumps({
-            "type": "geojson",
-            "url": "http://localhost:5000/outputs/MEDUSA_Roads.geojson",
-            "name": "MEDUSA Road Surfaces",
-            "id": "some unique I111D2"
+            "url": "http://localhost:8088/geoserver/db-pollution/ows",
+            "typeNames": "db-pollution:medusa2_model_output_buildings",
+            "parameters": {
+                "cql_filter": f"scenario_id={scenario_id}",
+            },
+            "maxFeatures": 300000
         })
 
+        response.outputs['roads'].data = json.dumps({
+            "type": "wfs",
+            "name": "MEDUSA Road Surfaces",
+            "url": "http://localhost:8088/geoserver/db-pollution/ows",
+            "typeNames": "db-pollution:medusa2_model_output_roads",
+            "parameters": {
+                "cql_filter": f"scenario_id={scenario_id}",
+            },
+            "maxFeatures": 10000
+        })
 
 # run_medusa_2.main(
 #     selected_polygon_gdf=sample_polygon,
