### 2024-10-23 5a1ab69
Remove status_supported, since it does not appear to be used within terria
