### 2024-10-23 d0de992
Nicer formatted medusa outputs variables
