### 2024-10-25 c948add
- Changed TSS, TCu, and TZn. In that, the k parameter affecting the positivity and negativity of the results. Here, I kept k as negative within `math.exp()` function to have positive results. - In the spreadsheet, Francis used calculated k for each Copper and Zinc roofs, constant k for TSS. Also, she used negative k in `EXP()` function. This is regardless of new or old formulas.
