+++ src/pollution_model/run_medusa_2.py
@@ -137,7 +137,7 @@ def compute_tss_roof_road(surface_area: float,
     Returns
     -------
     float
-       Returns the TSS value from the given parameters
+       Returns the TSS value from the given parameters (grams)
 
     Raises
     ----------
@@ -154,25 +154,25 @@ def compute_tss_roof_road(surface_area: float,
 
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.GALVANISED:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_OTHER:
             # Using coefficients of Galvanised rather than Copper
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_TILE:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.NON_METAL:
-            a1, a2, a3 = 0.6, 0.25, 0.008
+            a1, a2, k = 0.6, 0.25, 0.00933
         case SurfaceType.ZINCALUME:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.ASPHALT_ROAD | SurfaceType.CAR_PARK:
-            a1, a2, a3 = 190, 0.16, 0.0125
+            a1, a2, k = 190, 0.16, 0.00933
         case _:
             raise ValueError(invalid_surface_error)
     antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
-    first_term = surface_area * a1 * antecedent_dry_days ** a2 * capacity_factor
-    second_term = 1 - math.exp(a3 * average_rain_intensity * event_duration)
+    first_term = a1 * (antecedent_dry_days ** a2) * surface_area * capacity_factor
+    second_term = 1 - math.exp(-k * average_rain_intensity * event_duration)
 
     return first_term * second_term
 
@@ -243,16 +243,16 @@ def total_metal_load_roof(surface_area: float,
     # Define constants in a list
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
-            b = [2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
-            c = [910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
+            b = [None, 2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
+            c = [None, 910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
             # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
             k = 0.00933
             # Duration of the event measured by hours - Roof
             # observed from the intra-event concentration sampling
             z = 0.75
         case SurfaceType.GALVANISED:
-            b = [2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
-            c = [910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
+            b = [None, 2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
+            c = [None, 910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
             # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
             k = 0.00933
[…]
