### 2024-10-31 87e1305
move validate_aws_env_vars() to S3Manager class and update run_all.py accordingly
