+++ src/run_all.py
@@ -42,7 +42,7 @@ def main(
         - LogLevel.NOTSET (0)
     """
     # Validate that necessary AWS environment variables are set when S3 usage is enabled
-    s3_connection.validate_aws_env_vars()
+    s3_connection.S3Manager.validate_aws_env_vars()
     # Iterate through the dictionary containing modules and their parameters
     for module, parameters in modules_to_parameters.items():
         # Call the main function of each module with the selected polygon and specified parameters
