### 2024-11-01 #7266 @pritamd47
> @pritamd47 Thanks for agreeing to review this submission and welcome to the JOSS community! I will open a new issue with detailed instructions for the review process.

Hi @cheginit. I didn't receive any notification related to the issue you mention with instructions the review.
