### 2024-11-01 adcdc09
NEW FORMULAS - vers 001 - TSS-Roof: Change a3 into -k - TSS-Road: Change a3 into -k, change Cf (capacity) into (a7*INT) - TCu-Roof, TZn-Roof: Change constant k into negative calculated k and added 1000. - DCu-Roof, DZn-Roof: Change f to l and g to m. Change f=0.5 into l=0.77 in MetalOther and NoneMetal. Change g=0.43 into m=0.717 in MetalOther. - HAVE NOT CHANGE THE UNIT YET.
