+++ src/pollution_model/run_medusa_2.py
@@ -144,38 +144,49 @@ def compute_tss_roof_road(surface_area: float,
     ValueError
         If the surface type is not a roof or road (concrete, copper, galvanised roof, asphalt road, or car park).
     """
-    # Define the constants (Cf is the capacity factor).
-    capacity_factor = 0.75 if surface_type in ROOF_SURFACE_TYPES else 0.25
-    # Values a1 to a3 are empirically derived coefficient values.
-
     # Define error message (if needed)
     invalid_surface_error = (f"Given surface is not valid for computing TSS."
                              f" Needed a roof or road, but got {SurfaceType(surface_type).name}.")
 
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
+            # Galvanised Painted
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.GALVANISED:
+            # Galvanised
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_OTHER:
+            # Copper Old
             # Using coefficients of Galvanised rather than Copper
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_TILE:
+            # Decramastic
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.NON_METAL:
+            # Concrete, Concrete Painted, and Glass Roof
             a1, a2, k = 0.6, 0.25, 0.00933
         case SurfaceType.ZINCALUME:
+            # Zincalume
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.ASPHALT_ROAD | SurfaceType.CAR_PARK:
-            a1, a2, k = 190, 0.16, 0.00933
+            a1, a2, a7, k = 190, 0.16, 0.0125, 0.0008
         case _:
             raise ValueError(invalid_surface_error)
-    antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
-    first_term = a1 * (antecedent_dry_days ** a2) * surface_area * capacity_factor
-    # THE K WAS KEPT NEGATIVE IN THE MATH.EXP() FOR HAVING POSITIVE RESULTS
-    second_term = 1 - math.exp(-k * average_rain_intensity * event_duration)
 
-    return first_term * second_term
+    # Call out necessary variables
+    antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
+    # THE K WAS KEPT NEGATIVE IN THE MATH.EXP() FOR HAVING POSITIVE RESULTS
+    both_factor = 1 - math.exp(-k * average_rain_intensity * event_duration)
+
+    # NEED TO CHECK THE NAME OF ASPHALT ROAD AGAIN
+    if surface_type != 'AsphaltRoad':
+        # Roof
+        tss_factor = a1 * (antecedent_dry_days ** a2) * surface_area * 0.75
+    else:
+        # Road
+        tss_factor = a1 * (antecedent_dry_days ** a2) * surface_area * a7 * average_rain_intensity
+
+    return tss_factor * both_factor
 
 
 def total_metal_load_surface(surface_area: float,
@@ -241,6 +252,7 @@ def total_metal_load_roof(surface_area: float,
     # Define error message (if needed)
     invalid_surface_error = (f"Given surface is not valid for computing total metal load."
                              f" Needed a roof, but got {SurfaceType(surface_type).name}.")
+
     # Define constants in a list
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
@@ -311,6 +323,8 @@ def total_metal_load_roof(surface_area: float,
[…]
