### 2024-11-04 cdcfed2
Add Dockerfile for terriajs, so catalogs can be built into image instead of added through volumes
