+++ terriajs/Dockerfile
@@ -1,0 +1,5 @@
+FROM ghcr.io/terriajs/terriamap:0.2.1
+
+COPY client_config.json wwwroot/config.json
+COPY catalog.json wwwroot/init/catalog.json
+
