### 2024-11-06 42e2615
update get_elevations_near_rec_entry_point() in response to updated get_dem_band_and_resolution_by_geometry() in 'newzealidar'
