### 2024-11-06 #7433 @editorialbot
[REVIEW]: Flood Resilience Digital Twin: A system to enable flood risk analysis and prediction
**Submitting author:** @LukeParky (<a href="http://orcid.org/0000-0003-3130-3445">Luke Parkinson</a>)
**Repository:** https://github.com/GeospatialResearch/Digital-Twins
**Branch with paper.md** (empty if default branch): master
**Version:** v1.2.0
