+++ src/pollution_model/run_medusa_2.py
@@ -157,7 +157,8 @@ def compute_tss_roof_road(surface_area: float,
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_OTHER:
             # Copper Old
-            # Using coefficients of Galvanised rather than Copper
+            # Using coefficients of Galvanised rather than Copper as suggested by Francis
+            # Email named 'Aligning CCC Surface Type data with Okeover
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_TILE:
             # Decramastic
@@ -175,16 +176,16 @@ def compute_tss_roof_road(surface_area: float,
 
     # Call out necessary variables
     antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
-    # THE K WAS KEPT NEGATIVE IN THE MATH.EXP() FOR HAVING POSITIVE RESULTS
+    # Factor appears in both road and roof calculation
     both_factor = 1 - math.exp(-k * average_rain_intensity * event_duration)
 
-    # NEED TO CHECK THE NAME OF ASPHALT ROAD AGAIN
-    if surface_type != 'AsphaltRoad':
-        # Roof
-        tss_factor = a1 * (antecedent_dry_days ** a2) * surface_area * 0.75
-    else:
+    # Calculate tss for road and roof
+    if surface_type == 'AsphaltRoad':
         # Road
         tss_factor = a1 * (antecedent_dry_days ** a2) * surface_area * a7 * average_rain_intensity
+    else:
+        # Roof
+        tss_factor = a1 * (antecedent_dry_days ** a2) * surface_area * 0.75
 
     return tss_factor * both_factor * 1000
 
@@ -256,57 +257,46 @@ def total_metal_load_roof(surface_area: float,
     # Define constants in a list
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
-            # None was added to avoid the 0 position when slicing
+            # None was added to avoid the 0 position when indexing
             b = [None, 2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
             c = [None, 910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
-            # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
-            k = 0.00933
             # Duration of the event measured by hours - Roof
             # observed from the intra-event concentration sampling
             z = 0.75
         case SurfaceType.GALVANISED:
-            # None was added to avoid the 0 position when slicing
+            # None was added to avoid the 0 position when indexing
             b = [None, 2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
             c = [None, 910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
-            # Roof wash off coefficient (based on rate of decay to second concentrations from initial ones)
-            k = 0.00933
             # Duration of the event measured by hours - Roof
             # observed from the intra-event concentration sampling
             z = 0.75
         case SurfaceType.METAL_OTHER:
-            # None was added to avoid the 0 position when slicing
-            # Using coefficients of Galvanised rather than Copper
+            # None was added to avoid the 0 position when indexing
+            # Using coefficients of Galvanised rather than Copper as suggested by Frances
+            # Email named 'Aligning CCC Surface Type data with Okeover
             b = [None, 2, -2.802, 0.5, 0.217, 3.57, -0.09, 7, -3.732]
             c = [None, 910, 4, 0.2, 0.09, 1.5, -2, -0.23, 1.99]
[…]
