### 2024-11-06 0802b79
Change `both_factor` to `first_term` and `second_term`
