### 2024-11-06 9269d5f
Create test_run_medusa_2.py stubs
