### 2024-11-07 6761560
- Add 6 tests: 3 for roof and 3 for road
