### 2024-11-08 e63d520
- Add a test to check the total calculations
