### 2024-11-08 GeospatialResearch/Digital-Twins#279 @pritamd47
Reviewing for JOSS: Update LICENSE
Hi. I am a reviewer for your submission at JOSS at [@openjournals/joss-reviews#743](https://github.com/openjournals/joss-reviews/issues/7433). As per the recommendation of the journal's reviewing guidelines, I will be opening issues and relevant PRs if needed throughout the review. Please don't hesitate to reach out for clarification or help!

I notice that the LICENSE file has some additional text at the end that could be removed to make the language of the license as per the [OSI approved GNU license](https://opensource.org/license/agpl-v3). I think the LICENSE file should only have the lines until "END OF TERMS AND CONDITIONS". The lines under it should be added to each file of the code. See the end of the document at  https://opensource.org/license/agpl-v3 for more details.
