### 2024-11-12 e97dfe1
Freeze conda-pack because new versions cause side effects with scrapy imports.
