### 2024-11-12 a7f4e06
Include PyWPS required directories in Dockerfile
