### 2024-11-13 ec5e859
Run PyWPS processes within celery
