+++ src/tasks.py
@@ -20,7 +20,7 @@ from src.dynamic_boundary_conditions.rainfall import main_rainfall
 from src.dynamic_boundary_conditions.river import main_river
 from src.dynamic_boundary_conditions.tide import main_tide_slr
 from src.flood_model import bg_flood_model, process_hydro_dem
-from src.pollution_model.run_medusa_2 import retrieve_input_parameters
+from src.pollution_model import run_medusa_2
 from src.environmental.water_quality import surface_water_sites
 from src.run_all import DEFAULT_MODULES_TO_PARAMETERS
 
@@ -73,6 +73,22 @@ class DepthTimePlot(NamedTuple):
 
     depths: List[float]
     times: List[float]
+
+
+@app.task(base=OnFailureStateTask)
+def run_medusa_model(selected_polygon_wkt: str,
+                     antecedent_dry_days: float,
+                     average_rain_intensity: float,
+                     event_duration: float,
+                     rainfall_ph: float = 6.5) -> int:
+    selected_polygon = wkt_to_gdf(selected_polygon_wkt)
+    log_level = DEFAULT_MODULES_TO_PARAMETERS[run_medusa_2]["log_level"]
+    return run_medusa_2.main(selected_polygon,
+                      log_level,
+                      antecedent_dry_days,
+                      average_rain_intensity,
+                      event_duration,
+                      rainfall_ph)
 
 
 def create_model_for_area(selected_polygon_wkt: str, scenario_options: dict) -> result.GroupResult:
@@ -304,7 +320,7 @@ def retrieve_medusa_input_parameters(scenario_id: int) -> Optional[Dict[str, Uni
         A dictionary contain information from Rainfall MEDUSA 2.0 database or None if scenario does not exist.
     """
     # Get rainfall information
-    medusa_rainfall_event = retrieve_input_parameters(scenario_id)
+    medusa_rainfall_event = run_medusa_2.retrieve_input_parameters(scenario_id)
 
     # Return a dictionary format of these parameters or None if no ID found
     if medusa_rainfall_event is None:
