### 2024-11-13 d1ef095
view the input parameters used for a scenario run from the frontend
