### 2024-11-13 1795d77
Initialise DB when celery starts
