### 2024-11-13 ab33438
Move pollution model initialisation to retrieve_static_boundaries.py
