+++ src/digitaltwin/data_to_db.py
@@ -12,6 +12,7 @@ import pandas as pd
 from sqlalchemy.engine import Engine
 from sqlalchemy.sql import text
 
+from src.config import EnvVariable
 from src.digitaltwin.tables import GeospatialLayers, UserLogInfo, create_table, check_table_exists, execute_query
 from src.digitaltwin.get_data_using_geoapis import fetch_vector_data_using_geoapis
 
@@ -99,11 +100,11 @@ def get_geospatial_layer_info(layer_row: pd.Series) -> Tuple[str, int, str, str]
 
 
 def get_vector_data_id_not_in_db(
-        engine: Engine,
-        vector_data: gpd.GeoDataFrame,
-        table_name: str,
-        unique_column_name: str,
-        area_of_interest: gpd.GeoDataFrame) -> Set[int]:
+    engine: Engine,
+    vector_data: gpd.GeoDataFrame,
+    table_name: str,
+    unique_column_name: str,
+    area_of_interest: gpd.GeoDataFrame) -> Set[int]:
     """
     Get the IDs from the fetched vector_data that are not present in the specified database table
     for the area of interest.
@@ -146,9 +147,9 @@ def get_vector_data_id_not_in_db(
 
 
 def nz_geospatial_layers_data_to_db(
-        engine: Engine,
-        crs: int = 2193,
-        verbose: bool = False) -> None:
+    engine: Engine,
+    crs: int = 2193,
+    verbose: bool = False) -> None:
     """
     Fetch New Zealand geospatial layers data using 'geoapis' and store it into the database.
 
@@ -182,9 +183,9 @@ def nz_geospatial_layers_data_to_db(
 
 
 def get_non_intersection_area_from_db(
-        engine: Engine,
-        catchment_area: gpd.GeoDataFrame,
-        table_name: str) -> gpd.GeoDataFrame:
+    engine: Engine,
+    catchment_area: gpd.GeoDataFrame,
+    table_name: str) -> gpd.GeoDataFrame:
     """
     Get the non-intersecting area from the catchment area and user log information table in the database
     for the specified table.
@@ -241,13 +242,13 @@ def get_non_intersection_area_from_db(
 
 
 def process_new_non_nz_geospatial_layers(
-        engine: Engine,
-        data_provider: str,
-        layer_id: int,
-        table_name: str,
-        area_of_interest: gpd.GeoDataFrame,
-        crs: int = 2193,
-        verbose: bool = False) -> None:
+    engine: Engine,
+    data_provider: str,
+    layer_id: int,
+    table_name: str,
+    area_of_interest: gpd.GeoDataFrame,
+    crs: int = 2193,
+    verbose: bool = False) -> None:
     """
     Fetch new non-NZ geospatial layers data using 'geoapis' and store it into the database.
 
@@ -281,14 +282,14 @@ def process_new_non_nz_geospatial_layers(
 
 
 def process_existing_non_nz_geospatial_layers(
-        engine: Engine,
-        data_provider: str,
-        layer_id: int,
-        table_name: str,
-        unique_column_name: str,
-        area_of_interest: gpd.GeoDataFrame,
-        crs: int = 2193,
-        verbose: bool = False) -> None:
+    engine: Engine,
+    data_provider: str,
+    layer_id: int,
+    table_name: str,
+    unique_column_name: str,
+    area_of_interest: gpd.GeoDataFrame,
+    crs: int = 2193,
+    verbose: bool = False) -> None:
     """
     Fetch existing non-NZ geospatial layers data using 'geoapis' and store it into the database.
 
@@ -334,10 +335,10 @@ def process_existing_non_nz_geospatial_layers(
 
 
[…]
