### 2024-11-13 e1b3bad
Add roof surface data to docker image (pending approval before adding to public git repo)
