+++ Dockerfile
@@ -49,6 +49,7 @@ RUN mkdir /stored_data \
     && setfacl -R -m u:nonroot:rwx /stored_data \
     && mkdir /stored_data/geoserver
 
+# Add clipped roof surfaces data into image
 COPY --chown=root:root --chmod=444 \
     ./roof_surfaces_data/clipped_CCC_Lynker_RoofMaterials_Update_2023.gdb/ \
     /stored_data/roof_surfaces_data/roof_surfaces.gdb/
@@ -84,6 +85,7 @@ USER nonroot
 EXPOSE 5000
 
 SHELL ["/bin/bash", "-c"]
+# Run production HTTP server with threads (gevent) with an extended timeout
 ENTRYPOINT source /venv/bin/activate && \
            gunicorn --bind 0.0.0.0:5000 --timeout 600 -k gevent  src.app:app
 
