### 2024-11-13 61dfa79
- Add a test to check the total calculations and edit according to the comments
