+++ src/pollution_model/medusa_process_service.py
@@ -53,7 +53,6 @@ class MedusaProcessService(Process):
         response : ExecuteResponse
             The WPS response, containing output data.
         """
-
         # Helper function to format `number` for visualization
         def format_number(number: float) -> Union[int, float]:
             """Return `number` as an int if whole, otherwise as a float."""
