### 2024-11-13 2e27723
rename format_number() to _format_number()
