### 2024-11-13 9cdc577
quick update medusa handler
