### 2024-11-14 eacec79
Remove Cesium front-end
