+++ visualisation/env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="vite/client" />
-// Used to ensure that vite builds have the correct type declarations.
