### 2024-11-14 9245e8a
Change `assertEquals` to `assertEqual`
