### 2024-11-17 2724736
Revert "Move pollution model initialisation to retrieve_static_boundaries.py"
This reverts commit ab33438f
