+++ src/digitaltwin/tables.py
@@ -4,12 +4,12 @@
 from datetime import datetime, timezone
 
 from geoalchemy2 import Geometry
-from sqlalchemy import Boolean, Column, DateTime, inspect, Integer, String
+from sqlalchemy import Column, DateTime, inspect, Integer, String
 from sqlalchemy.dialects.postgresql import ARRAY
 from sqlalchemy.engine import Engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import Session, Query
-from sqlalchemy.schema import CheckConstraint, PrimaryKeyConstraint
+from sqlalchemy.schema import CheckConstraint
 
 Base = declarative_base()
 
@@ -79,120 +79,6 @@ class UserLogInfo(Base):
     source_table_list = Column(ARRAY(String), comment="associated tables (geospatial layers)")
     created_at = Column(DateTime(timezone=True), default=datetime.now(timezone.utc), comment="log created datetime")
     geometry = Column(Geometry("POLYGON", srid=2193))
-
-
-class RiverNetworkExclusions(Base):
-    """
-    Class representing the 'rec_network_exclusions' table.
-
-    Attributes
-    ----------
-    __tablename__ : str
-        Name of the database table.
-    rec_network_id : int
-        An identifier for the river network associated with each new run.
-    objectid : int
-        An identifier for the REC river object matching from the 'rec_data' table.
-    exclusion_cause : str
-        Cause of exclusion, i.e., the reason why the REC river geometry was excluded.
-    geometry : LineString
-        Geometric representation of the excluded REC river features.
-    """  # pylint: disable=too-few-public-methods
-
-    __tablename__ = "rec_network_exclusions"
-    rec_network_id = Column(Integer, primary_key=True,
-                            comment="An identifier for the river network associated with each run")
-    objectid = Column(Integer, primary_key=True,
-                      comment="An identifier for the REC river object matching from the 'rec_data' table")
-    exclusion_cause = Column(String, comment="Cause of exclusion")
-    geometry = Column(Geometry("LINESTRING", srid=2193))
-
-    __table_args__ = (
-        PrimaryKeyConstraint('rec_network_id', 'objectid', name='network_exclusions_pk'),
-    )
-
-
-class RiverNetwork(Base):
-    """
-    Class representing the 'rec_network' table.
-
-    Attributes
-    ----------
-    __tablename__ : str
-        Name of the database table.
-    rec_network_id : int
-        An identifier for the river network associated with each new run (primary key).
-    network_path : str
-        Path to the REC river network file.
-    network_data_path : str
-        Path to the REC river network data file for the AOI.
-    created_at : datetime
-        Timestamp indicating when the output was created.
-    geometry : Polygon
-        Geometric representation of the catchment area coverage.
-    """  # pylint: disable=too-few-public-methods
-
-    __tablename__ = "rec_network"
-    rec_network_id = Column(Integer, primary_key=True,
-                            comment="An identifier for the river network associated with each run")
-    network_path = Column(String, comment="path to the rec river network file")
-    network_data_path = Column(String, comment="path to the rec river network data file for the AOI")
-    created_at = Column(DateTime(timezone=True), default=datetime.now(timezone.utc), comment="output created datetime")
-    geometry = Column(Geometry("POLYGON", srid=2193))
-
-
-class BGFloodModelOutput(Base):
-    """
-    Class representing the 'bg_flood_model_output' table.
-
-    Attributes
-    ----------
-    __tablename__ : str
-        Name of the database table.
-    unique_id : int
-        Unique identifier for each entry (primary key).
-    file_name : str
-        Name of the flood model output file.
-    file_path : str
-        Path to the flood model output file.
-    created_at : datetime
[…]
