### 2024-11-17 fc67062
View the input parameters used for a scenario run from the frontend (#284)
* reorder catalog.json medusa traits

* view the input parameters used for a scenario run from the frontend

* fix lint

* fix lint

* fix lint

* rename format_number() to _format_number()

* quick update medusa handler
