+++ src/pollution_model/medusa_process_service.py
@@ -1,8 +1,11 @@
+# -*- coding: utf-8 -*-
 """Defines PyWPS WebProcessingService process for running MEDUSA model"""
+
 import json
+from typing import Union
 
 import geopandas as gpd
-from pywps import ComplexOutput, Format, LiteralInput, Process, WPSRequest
+from pywps import ComplexOutput, Format, LiteralInput, LiteralOutput, Process, WPSRequest
 from pywps.inout.literaltypes import AnyValue
 from pywps.response.execute import ExecuteResponse
 
@@ -15,15 +18,20 @@ class MedusaProcessService(Process):
     def __init__(self) -> None:
         """Define inputs and outputs of the WPS process, and assign process handler."""
         inputs = [
-            LiteralInput("antecedentDryDays", "Antecedent Dry Days", data_type='float', allowed_values=AnyValue()),
-            LiteralInput("averageRainIntensity", "Average Rain Intensity (mm/hour)", data_type='float',
-                         allowed_values=AnyValue()),
-            LiteralInput("eventDuration", "Event Duration (hours)", data_type='float', allowed_values=AnyValue()),
+            LiteralInput("antecedentDryDays", "Antecedent Dry Days",
+                         data_type='float', allowed_values=AnyValue()),
+            LiteralInput("averageRainIntensity", "Average Rain Intensity (mm/hour)",
+                         data_type='float', allowed_values=AnyValue()),
+            LiteralInput("eventDuration", "Event Duration (hours)",
+                         data_type='float', allowed_values=AnyValue()),
         ]
         outputs = [
+            LiteralOutput("scenarioDetails", "Scenario Details",
+                          data_type='string'),
             ComplexOutput("roofs", "Output",
                           supported_formats=[Format("application/vnd.terriajs.catalog-member+json")]),
-            ComplexOutput("roads", "Output", supported_formats=[Format("application/vnd.terriajs.catalog-member+json")])
+            ComplexOutput("roads", "Output",
+                          supported_formats=[Format("application/vnd.terriajs.catalog-member+json")])
         ]
         super().__init__(
             self._handler,
@@ -45,6 +53,23 @@ class MedusaProcessService(Process):
         response : ExecuteResponse
             The WPS response, containing output data.
         """
+        # Helper function to format `number` for visualization
+        def _format_number(number: float) -> Union[int, float]:
+            """
+            Return `number` as an int if whole, otherwise as a float.
+
+            Parameters
+            ----------
+            number : float
+                The number to be formatted.
+
+            Returns
+            -------
+            Union[int, float]
+                The formatted number, either as an int if whole or as a float.
+            """
+            return int(number) if number % 1 == 0 else float(number)
+
         # Read input parameters from request
         antecedent_dry_days = request.inputs['antecedentDryDays'][0].data
         average_rain_intensity = request.inputs['averageRainIntensity'][0].data
@@ -59,6 +84,26 @@ class MedusaProcessService(Process):
         # Wait until celery task is completed
         scenario_id = medusa_task.get()
 
+        # Create scenario details as HTML-formatted text with input values and scenario ID
+        scenario_details = (
+            f"Antecedent Dry Days: {_format_number(antecedent_dry_days)}<br>"
+            f"Average Rain Intensity (mm/hour): {_format_number(average_rain_intensity)}<br>"
+            f"Event Duration (hours): {_format_number(event_duration)}<br>"
+            f"Scenario ID: {_format_number(scenario_id)}"
+        )
+
+        # Create a short report containing scenario details
+        scenario_short_report = [
+            {
+                "name": "Scenario Details",
+                "content": scenario_details,
+                "show": False
+            }
+        ]
+
[…]
