### 2024-11-17 ee8f88d
Remove linting for javascript, since all javascript has been removed
