+++ .github/workflows/run-tests.yml
@@ -48,38 +48,6 @@ jobs:
         run: |
           pylint src
 
-  linting-js:
-    name: Lint Front-end application
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./visualisation
-    steps:
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 16.x
-      - uses: actions/checkout@v3
-      - name: Install dependencies
-        run: npm ci
-      - name: Run linting
-        run: npm run lint
-
-  type-check-ts:
-    name: Type check TypeScript front-end
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./visualisation
-    steps:
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 16.x
-      - uses: actions/checkout@v3
-      - name: Install dependencies
-        run: npm ci
-      - name: Run type check
-        run: npm run type-check
-
   unit-tests:
     name: Run unit tests with pytest
     runs-on: ubuntu-latest
