### 2024-11-17 e21e6f4
Update medusa tests to cover multiple different surface types
