### 2024-11-17 93cf0ca
276 fix medusa output values being incorrect (#277)
Fix incorrect outputs from MEDUSA, add unit tests.
* - Changed TSS, TCu, and TZn. In that, the k parameter affecting the positivity and negativity of the results. Here, I kept k as negative within `math.exp()` function to have positive results.
- In the spreadsheet, Francis used calculated k for each Copper and Zinc roofs, constant k for TSS. Also, she used negative k in `EXP()` function. This is regardless of new or old formulas.

* - Adding comments

* NEW FORMULAS - vers 001
- TSS-Roof: Change a3 into -k
- TSS-Road: Change a3 into -k, chan
[…]
