+++ src/pollution_model/run_medusa_2.py
@@ -137,44 +137,58 @@ def compute_tss_roof_road(surface_area: float,
     Returns
     -------
     float
-       Returns the TSS value from the given parameters
+       Returns the TSS value from the given parameters (micrograms)
 
     Raises
     ----------
     ValueError
         If the surface type is not a roof or road (concrete, copper, galvanised roof, asphalt road, or car park).
     """
-    # Define the constants (Cf is the capacity factor).
-    capacity_factor = 0.75 if surface_type in ROOF_SURFACE_TYPES else 0.25
-    # Values a1 to a3 are empirically derived coefficient values.
-
     # Define error message (if needed)
     invalid_surface_error = (f"Given surface is not valid for computing TSS."
                              f" Needed a roof or road, but got {SurfaceType(surface_type).name}.")
 
     match surface_type:
         case SurfaceType.COLOUR_STEEL:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            # Galvanised Painted
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.GALVANISED:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            # Galvanised
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_OTHER:
-            # Using coefficients of Galvanised rather than Copper
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            # Copper Old
+            # Using coefficients of Galvanised rather than Copper as suggested by Francis
+            # Email named 'Aligning CCC Surface Type data with Okeover
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.METAL_TILE:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            # Decramastic
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.NON_METAL:
-            a1, a2, a3 = 0.6, 0.25, 0.008
+            # Concrete, Concrete Painted, and Glass Roof
+            a1, a2, k = 0.6, 0.25, 0.00933
         case SurfaceType.ZINCALUME:
-            a1, a2, a3 = 0.4, 0.5, 0.008
+            # Zincalume
+            a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.ASPHALT_ROAD | SurfaceType.CAR_PARK:
-            a1, a2, a3 = 190, 0.16, 0.0125
+            a1, a2, k = 190, 0.16, 0.0008
         case _:
             raise ValueError(invalid_surface_error)
-    antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
-    first_term = surface_area * a1 * antecedent_dry_days ** a2 * capacity_factor
-    second_term = 1 - math.exp(a3 * average_rain_intensity * event_duration)
 
-    return first_term * second_term
+    # Call out necessary variables
+    antecedent_dry_days, average_rain_intensity, event_duration, _ph = rainfall_event
+
+    # Calculate tss for road and roof
+    if surface_type == 'AsphaltRoad':
+        # Road
+        first_term = a1 * (antecedent_dry_days ** a2) * surface_area * 0.25
+    else:
+        # Roof
+        first_term = a1 * (antecedent_dry_days ** a2) * surface_area * 0.75
+
+    # Second term appears in both road and roof calculation
+    second_term = 1 - math.exp(-k * average_rain_intensity * event_duration)
+
+    return first_term * second_term * 1000
 
 
 def total_metal_load_surface(surface_area: float,
@@ -240,84 +254,100 @@ def total_metal_load_roof(surface_area: float,
     # Define error message (if needed)
[…]
