+++ src/pywps.cfg
@@ -1,25 +1,21 @@
 [metadata:main]
-identification_title=PyWPS Demo server
-identification_abstract=PyWPS testing and development server. Do NOT use this server in production environment. You shall setup PyWPS as WSGI application for production. Please refer documentation for further detials.
-identification_keywords=WPS,GRASS,PyWPS, Demo, Dev
+identification_title=Ōtākaro Digital Twin
+identification_abstract=WPS services for the Digital Twin for the Ōtākaro Avon River, in Christchurch New Zealand.
+identification_keywords=WPS, PyWPS, DigitalTwin, Otakaro, Avon
 identification_keywords_type=theme
 identification_fees=None
 identification_accessconstraints=None
-provider_name=PyWPS Developement team
-provider_url=http://pywps.org/'
-contact_name=Your Name
-contact_position=Developer
-contact_address=My Street
-contact_city=My City
+provider_name=GeospatialResearchInstitute
+provider_url=https://geospatial.ac.nz
+contact_name=Luke Parkinson
+contact_position=Geospatial Software Developer
+contact_address=University of Canterbury
+contact_city=Christchurch
 contact_stateorprovince=None
-contact_postalcode=000 00
-contact_country=World, Internet
-contact_phone=+00 00 11 22 33
-contact_fax=+00 99 88 77 66
-contact_email=info@yourdomain.org
-contact_url=http://pywps.org
-contact_hours=8:00-20:00UTC
-contact_instructions=Knock on the door
+contact_postalcode=8140
+contact_country=New Zealand
+contact_email=luke.parkinson@canterbury.ac.nz
+contact_instructions=Contact via Email or GitHub
 contact_role=pointOfContact
 
 [server]
