### 2024-11-18 f823d19
Apply suggestions from code review
Co-authored-by: Casey <casey.li@canterbury.ac.nz>
