+++ src/digitaltwin/retrieve_static_boundaries.py
@@ -45,7 +45,7 @@ def main(
     # Store user log information in the database
     data_to_db.user_log_info_to_db(engine, catchment_area)
     # Read roof_surface file and store in the database
-    data_to_db.save_roof_surface_type_data_to_db(engine)
+    data_to_db.save_roof_surface_data_to_db(engine)
 
 
 if __name__ == "__main__":
