### 2024-11-18 e3d277e
Add environment variable controls for pywps.cfg
