+++ src/pywps.cfg
@@ -21,8 +21,8 @@ contact_role=pointOfContact
 [server]
 maxsingleinputsize=1mb
 maxrequestsize=3mb
-url=http://localhost:5000/wps
-outputurl=http://localhost:5000/outputs/
+url=$BACKEND_HOST:$BACKEND_PORT/wps
+outputurl=$BACKEND_HOST:$BACKEND_PORT/outputs/
 outputpath=outputs
 workdir=workdir
 wd_inp_subdir=inputs
