### 2024-11-18 5261998
Add environment variable substitution for terriajs
