### 2024-11-18 dac56d1
280 remove original geo visualisation components front end (#288)
* Remove Cesium front-end

* Update dcumentation to reference terria

* Remove linting for javascript, since all javascript has been removed
