+++ docker-compose.yml
@@ -96,25 +96,6 @@ services:
       timeout: 5s
       retries: 10
 
-  www:
-    # Webserver for the website interface
-    image: lparkinson/www-flood-resilience-dt:1.2
-    build:
-      context: .
-      dockerfile: visualisation/Dockerfile
-    container_name: www_digital_twin
-    environment:
-      - VITE_CESIUM_ACCESS_TOKEN=$CESIUM_ACCESS_TOKEN
-      - VITE_GEOSERVER_HOST=$GEOSERVER_HOST
-      - VITE_GEOSERVER_PORT=$GEOSERVER_PORT
-      - VITE_POSTGRES_DB=$POSTGRES_DB
-    ports:
-      - "${WWW_PORT}:80"
-    healthcheck:
-      test: curl --fail -s http://localhost:80/ || exit 1
-      timeout: 5s
-      retries: 10
-
   terria_map:
     # Front-end using terria.js
     build:
@@ -127,7 +108,6 @@ services:
       test: curl --fail -s http://localhost:3001/ || exit 1
       timeout: 5s
       retries: 10
-
 
   message_broker:
     # Communicates between backend and workers to assign tasks and store state
