### 2024-11-18 a308910
Use single docker image for backend and celery-worker, but separate entrypoints
