### 2024-11-18 44e956f
Use single docker image for backend and celery-worker, but separate entrypoints (#292)
