### 2024-11-19 63ca97e
Simplify intersection SQL query by removing sub-query
