### 2024-11-19 4c6ac20
Solve problem with nz_roads and nz_building_outlines being fetched every time when running in celery (rounding error)
