### 2024-11-19 bddc88e
Remove use of GeoSeries.to_wkt() because it rounds too strongly and contributes to errors
