### 2024-11-19 cb7104f
Add WWW_PORT to terria config
