### 2024-11-19 afed040
Configure backend host and port in terria_map using environment variables
