### 2024-11-20 09d4f79
Configure geoserver url from env variable in PyWPS PyWPS
