### 2024-11-20 f7f11a1
289 Fix AOI intersection calculation problems (#293)
* Simplify intersection SQL query by removing sub-query

* Solve problem with nz_roads and nz_building_outlines being fetched every time when running in celery (rounding error)

* Introduce slight rounding errors into dev run_all to match production more closely

* Remove use of GeoSeries.to_wkt() because it rounds too strongly and contributes to errors
