+++ src/dynamic_boundary_conditions/river/river_network_to_from_db.py
@@ -146,7 +146,7 @@ def get_network_output_metadata(
     # Get the absolute path of the REC Network data file as a string
     network_data_path = network_data_path.as_posix()
     # Get the WKT representation of the catchment area's geometry
-    catchment_geom = catchment_area["geometry"].to_wkt().iloc[0]
+    catchment_geom = catchment_area.geometry[0].wkt
     # Return the metadata as a tuple
     return network_path, network_data_path, catchment_geom
 
