### 2024-11-20 d547e60
Use cleaner HEREDOC style Run scripts in terriajs/Dockerfile
