### 2024-11-20 ec55392
Configure read and write permissions on terria config
