### 2024-11-20 cb8a733
Use different local and docker geoserver URLs
