### 2024-11-20 9d27794
Prepare terria map for deployment (#294)
* Add WWW_PORT to terria config

* Configure backend host and port in terria_map using environment variables

* Configure geoserver url from env variable in PyWPS
PyWPS

* Remove geoserver and backend config from .env.docker-override, it is not needed

* Use cleaner HEREDOC style Run scripts in terriajs/Dockerfile

* Configure read and write permissions on terria config

* Use node instead of nonroot as user

* Use different local and docker geoserver URLs

* fix lint
