+++ docker-compose.yml
@@ -103,9 +103,11 @@ services:
     build:
       context: terriajs
     ports:
-      - "3001:3001"
+      - "${WWW_PORT}:3001"
     environment:
       - CESIUM_ACCESS_TOKEN=$CESIUM_ACCESS_TOKEN
+      - BACKEND_HOST=$BACKEND_HOST
+      - BACKEND_PORT=$BACKEND_PORT
     healthcheck:
       test: curl --fail -s http://localhost:3001/ || exit 1
       timeout: 5s
