### 2024-11-20 ecaab72
Group zinc outputs then zinc outputs, rather than total then dissolved
