### 2024-11-20 462f6c6
Display MEDUSA outputs to 5 significant figures
