+++ src/pollution_model/run_medusa_2.py
@@ -36,8 +36,8 @@ class SurfaceType(StrEnum):
 
     Attributes
     ----------
-    COLOUR_STEEL : str
-        Colour Steel surface
+    COLORSTEEL : str
+        Colorsteel surface
     GALVANISED : str
         Galvanised surface.
     METAL_OTHER : str
@@ -54,7 +54,7 @@ class SurfaceType(StrEnum):
         Car park.
     """
 
-    COLOUR_STEEL = "ColourSteel"
+    COLORSTEEL = "Colorsteel"
     GALVANISED = "Galvanised"
     METAL_OTHER = "MetalOther"
     METAL_TILE = "MetalTile"
@@ -64,7 +64,7 @@ class SurfaceType(StrEnum):
     CAR_PARK = "CarPark"  # CarParks are classified the same as roads
 
 
-ROOF_SURFACE_TYPES = {SurfaceType.COLOUR_STEEL, SurfaceType.GALVANISED, SurfaceType.METAL_OTHER, SurfaceType.METAL_TILE,
+ROOF_SURFACE_TYPES = {SurfaceType.COLORSTEEL, SurfaceType.GALVANISED, SurfaceType.METAL_OTHER, SurfaceType.METAL_TILE,
                       SurfaceType.NON_METAL, SurfaceType.ZINCALUME}
 ROAD_SURFACE_TYPES = {SurfaceType.ASPHALT_ROAD, SurfaceType.CAR_PARK}
 
@@ -110,9 +110,9 @@ class MetalLoads(NamedTuple):
     Attributes
     ----------
     cu_load: float
-        Amount of copper load contributed by a step in the pollutant model (micrograms).
+        Amount of copper load contributed by a step in the pollutant model (milligrams).
     zn_load: float
-        Amount of zinc load contributed by a step in the pollutant model (micrograms).
+        Amount of zinc load contributed by a step in the pollutant model (milligrams).
     """
 
     cu_load: float
@@ -137,7 +137,7 @@ def compute_tss_roof_road(surface_area: float,
     Returns
     -------
     float
-       Returns the TSS value from the given parameters (micrograms)
+       Returns the TSS value from the given parameters (milligrams)
 
     Raises
     ----------
@@ -149,7 +149,7 @@ def compute_tss_roof_road(surface_area: float,
                              f" Needed a roof or road, but got {SurfaceType(surface_type).name}.")
 
     match surface_type:
-        case SurfaceType.COLOUR_STEEL:
+        case SurfaceType.COLORSTEEL:
             # Galvanised Painted
             a1, a2, k = 0.4, 0.5, 0.00933
         case SurfaceType.GALVANISED:
@@ -212,7 +212,7 @@ def total_metal_load_surface(surface_area: float,
     Returns
     -------
     MetalLoads
-       Returns the total copper and zinc loads from the given parameters (micrograms).
+       Returns the total copper and zinc loads from the given parameters (milligrams).
 
     Raises
     ----------
@@ -244,7 +244,7 @@ def total_metal_load_roof(surface_area: float,
     Returns
     -------
     MetalLoads
-       Returns the total copper and zinc loads from the given parameters (micrograms).
+       Returns the total copper and zinc loads from the given parameters (milligrams).
 
     Raises
     ----------
@@ -257,7 +257,7 @@ def total_metal_load_roof(surface_area: float,
 
     # Define constants in a list
     match surface_type:
-        case SurfaceType.COLOUR_STEEL:
+        case SurfaceType.COLORSTEEL:
             # None was added to avoid the 0 position when indexing
[…]
