### 2024-11-21 2051c37
Split codebase by files into high-level core, floodresilience, otakaro
