+++ floodresilience/dynamic_boundary_conditions/river/hydrograph.py
@@ -10,7 +10,7 @@ import re
 
 import geopandas as gpd
 
-from src.dynamic_boundary_conditions.river.river_enum import BoundType
+from floodresilience.dynamic_boundary_conditions.river.river_enum import BoundType
 
 log = logging.getLogger(__name__)
 
