+++ tests/test_dynamic_boundary_conditions/rainfall/test_rainfall_model_input.py
@@ -5,7 +5,7 @@ import geopandas as gpd
 import pandas as pd
 import numpy as np
 
-from src.dynamic_boundary_conditions.rainfall import rainfall_model_input
+from floodresilience.dynamic_boundary_conditions.rainfall import rainfall_model_input
 
 
 class RainfallModelInputTest(unittest.TestCase):
@@ -63,7 +63,7 @@ class RainfallModelInputTest(unittest.TestCase):
         result = intersection_area_sizes.gt(org_area_sizes).any()
         self.assertFalse(result)
 
-    @patch("src.dynamic_boundary_conditions.rainfall.rainfall_model_input.sites_voronoi_intersect_catchment")
+    @patch("floodresilience.dynamic_boundary_conditions.rainfall.rainfall_model_input.sites_voronoi_intersect_catchment")
     def test_sites_coverage_in_catchment_correct_area_percent(self, mock_intersections):
         """Test to ensure the percentage of area covered by each rainfall site inside the catchment area has
         been calculated correctly and sums up to 1."""
