+++ floodresilience/dynamic_boundary_conditions/rainfall/rainfall_model_input.py
@@ -13,8 +13,8 @@ import pandas as pd
 import xarray as xr
 from geocube.api.core import make_geocube
 
-from src.dynamic_boundary_conditions.rainfall import hyetograph
-from src.dynamic_boundary_conditions.rainfall.rainfall_enum import RainInputType
+from floodresilience.dynamic_boundary_conditions.rainfall import hyetograph
+from floodresilience.dynamic_boundary_conditions.rainfall.rainfall_enum import RainInputType
 
 log = logging.getLogger(__name__)
 
