+++ floodresilience/dynamic_boundary_conditions/rainfall/main_rainfall.py
@@ -11,7 +11,7 @@ import geopandas as gpd
 from src import config
 from src.digitaltwin import setup_environment
 from src.digitaltwin.utils import LogLevel, setup_logging, get_catchment_area
-from src.dynamic_boundary_conditions.rainfall import (
+from floodresilience.dynamic_boundary_conditions.rainfall import (
     rainfall_sites,
     thiessen_polygons,
     hirds_rainfall_data_to_db,
@@ -19,7 +19,7 @@ from src.dynamic_boundary_conditions.rainfall import (
     hyetograph,
     rainfall_model_input,
 )
-from src.dynamic_boundary_conditions.rainfall.rainfall_enum import RainInputType, HyetoMethod
+from floodresilience.dynamic_boundary_conditions.rainfall.rainfall_enum import RainInputType, HyetoMethod
 
 
 def main(
