+++ tests/test_dynamic_boundary_conditions/rainfall/test_thiessen_polygons.py
@@ -2,7 +2,7 @@ import unittest
 
 import geopandas as gpd
 
-from src.dynamic_boundary_conditions.rainfall import thiessen_polygons
+from floodresilience.dynamic_boundary_conditions.rainfall import thiessen_polygons
 
 
 class ThiessenPolygonsTest(unittest.TestCase):
