### 2024-11-21 7476213
Split codebase by files into high-level core, floodresilience, otakaro
