+++ tests/test_dynamic_boundary_conditions/rainfall/test_rainfall_data_from_hirds.py
@@ -5,7 +5,7 @@ import math
 
 import pandas as pd
 
-from src.dynamic_boundary_conditions.rainfall import rainfall_data_from_hirds
+from floodresilience.dynamic_boundary_conditions.rainfall import rainfall_data_from_hirds
 
 
 class RainfallDataFromHirdsTest(unittest.TestCase):
