+++ floodresilience/dynamic_boundary_conditions/river/main_river.py
@@ -12,7 +12,7 @@ import geopandas as gpd
 from src import config
 from src.digitaltwin import setup_environment
 from src.digitaltwin.utils import LogLevel, setup_logging, get_catchment_area
-from src.dynamic_boundary_conditions.river import (
+from floodresilience.dynamic_boundary_conditions.river import (
     river_data_to_from_db,
     river_network_for_aoi,
     align_rec_osm,
@@ -20,7 +20,7 @@ from src.dynamic_boundary_conditions.river import (
     hydrograph,
     river_model_input
 )
-from src.dynamic_boundary_conditions.river.river_enum import BoundType
+from floodresilience.dynamic_boundary_conditions.river.river_enum import BoundType
 
 log = logging.getLogger(__name__)
 
