+++ floodresilience/dynamic_boundary_conditions/rainfall/hirds_rainfall_data_to_db.py
@@ -9,7 +9,7 @@ import geopandas as gpd
 from sqlalchemy.engine import Engine
 
 from src.digitaltwin import tables
-from src.dynamic_boundary_conditions.rainfall import rainfall_data_from_hirds
+from floodresilience.dynamic_boundary_conditions.rainfall import rainfall_data_from_hirds
 
 log = logging.getLogger(__name__)
 
