+++ floodresilience/dynamic_boundary_conditions/river/river_inflows.py
@@ -14,8 +14,8 @@ import xarray as xr
 from newzealidar.utils import get_dem_band_and_resolution_by_geometry
 from sqlalchemy.engine import Engine
 
-from src.dynamic_boundary_conditions.river import align_rec_osm
-from src.flood_model import process_hydro_dem
+from floodresilience.dynamic_boundary_conditions.river import align_rec_osm
+from floodresilience.flood_model import process_hydro_dem
 
 log = logging.getLogger(__name__)
 
