+++ floodresilience/dynamic_boundary_conditions/river/align_rec_osm.py
@@ -12,8 +12,8 @@ import pandas as pd
 from shapely.geometry import Point
 from sqlalchemy.engine import Engine
 
-from src.dynamic_boundary_conditions.river import osm_waterways
-from src.flood_model import process_hydro_dem
+from floodresilience.dynamic_boundary_conditions.river import osm_waterways
+from floodresilience.flood_model import process_hydro_dem
 
 log = logging.getLogger(__name__)
 
