### 2024-11-21 d01371f
Avoid use of inplace=True
