+++ src/app.py
@@ -2,74 +2,21 @@
 """The main web application that serves the Digital Twin to the web through a Rest API."""
 
 import logging
-import pathlib
-from functools import wraps
-from http.client import OK, ACCEPTED, BAD_REQUEST, INTERNAL_SERVER_ERROR, NOT_FOUND, SERVICE_UNAVAILABLE
-from typing import Callable, Dict, Tuple
+from http.client import OK
 
-import requests
-from celery import result, states
-from flask import Flask, Response, jsonify, make_response, send_file, request
+from flask import Flask, Response
 from flask_cors import CORS
 from flask_swagger_ui import get_swaggerui_blueprint
-from kombu.exceptions import OperationalError
 from pywps import Service
-from shapely import box
 
-from src import tasks
-from src.config import EnvVariable
+from floodresilience.blueprint import flood_resilience_blueprint
+from src.check_celery_alive import check_celery_alive
 from otakaro.pollution_model.medusa_process_service import MedusaProcessService
+from otakaro.blueprint import otakaro_blueprint
 
 # Initialise flask server object
 app = Flask(__name__)
 CORS(app)
-
-
-def check_celery_alive(f: Callable[..., Response]) -> Callable[..., Response]:
-    """
-    Check if the Celery workers are running and return INTERNAL_SERVER_ERROR if they are down using function decorator.
-
-    Parameters
-    ----------
-    f : Callable[..., Response]
-        The view function that is being decorated.
-
-    Returns
-    -------
-    Callable[..., Response]
-        Response is SERVICE_UNAVAILABLE if the celery workers are down, otherwise continue to function f
-    """
-
-    @wraps(f)
-    def decorated_function(*args: Tuple, **kwargs: Dict) -> Response:
-        """
-        Before function `f` is called, check if Celery workers are down, and return and error response if so.
-        If Celery workers are running, then continue with calling `f` with original arguments.
-
-        Parameters
-        ----------
-        args : Tuple
-            The original arguments for function `f`.
-        kwargs : Dict
-            The original keyword arguments for function `f`.
-
-        Returns
-        -------
-        Response
-            SERVICE_UNAVAILABLE if Celery workers are down, otherwise response from function `f`.
-        """
-        try:
-            ping_celery_response = tasks.app.control.ping()
-            if len(ping_celery_response) == 0:
-                logging.warning("Celery workers not active, may indicate a fault")
-                return make_response("Celery workers not active", SERVICE_UNAVAILABLE)
-        except OperationalError:
-            logging.warning("Celery workers not active, may indicate a fault")
-            return make_response("Celery workers not active", SERVICE_UNAVAILABLE)
-        return f(*args, **kwargs)
-
-    return decorated_function
-
 
 # Serve API documentation
 SWAGGER_URL = "/swagger"
@@ -80,6 +27,9 @@ swagger_ui_blueprint = get_swaggerui_blueprint(
     config={"app_name": "Flood Resilience Digital Twin (FReDT)"}
 )
 app.register_blueprint(swagger_ui_blueprint, url_prefix=SWAGGER_URL)
+
+app.register_blueprint(flood_resilience_blueprint)
+app.register_blueprint(otakaro_blueprint)
 
 processes = [
     MedusaProcessService()
@@ -108,20 +58,6 @@ def index() -> Response:
     """, OK)
 
 
-@app.route('/wps', methods=['GET', 'POST'])
-@check_celery_alive
-def wps() -> Service:
-    """
-    End point for OGC WebProcessingService spec, allowing clients such as TerriaJS to request processing.
-
-    Returns
-    -------
-    Service
-        The PyWPS WebProcessing Service instance
-    """
-    return service
-
-
[…]
