+++ floodresilience/blueprint.py
@@ -1,0 +1,307 @@
+"""Endpoints and flask configuration for the Flood Resilience Digital Twin"""
+
+from http.client import OK, ACCEPTED, BAD_REQUEST, INTERNAL_SERVER_ERROR, NOT_FOUND
+import pathlib
+
+from celery import result, states
+from flask import Blueprint, Response, jsonify, make_response, send_file, request
+import requests
+from shapely import box
+
+from floodresilience import tasks
+from src.check_celery_alive import check_celery_alive
+from src.config import EnvVariable
+
+flood_resilience_blueprint = Blueprint('floodresilience', __name__)
+
+
+@flood_resilience_blueprint.route('/tasks/<task_id>', methods=["GET"])
+def get_status(task_id: str) -> Response:
+    """
+    Retrieve status of a particular Celery backend task.
+    Supported methods: GET
+
+    Parameters
+    ----------
+    task_id : str
+        The id of the Celery task to retrieve status from
+
+    Returns
+    -------
+    Response
+        JSON response containing taskStatus
+    """
+    task_result = result.AsyncResult(task_id, app=tasks.app)
+    status = task_result.status
+    http_status = OK
+    if status == states.SUCCESS:
+        task_value = task_result.get()
+    elif status == states.FAILURE:
+        http_status = INTERNAL_SERVER_ERROR
+        is_debug_mode = EnvVariable.DEBUG_TRACEBACK
+        task_value = task_result.traceback if is_debug_mode else None
+    else:
+        task_value = None
+
+    return make_response(jsonify({
+        "taskId": task_result.id,
+        "taskStatus": status,
+        "taskValue": task_value
+    }), http_status)
+
+
+@flood_resilience_blueprint.route('/models/generate', methods=["POST"])
+@check_celery_alive
+def generate_model() -> Response:
+    """
+    Generate a flood model for a given area.
+    Supported methods: POST
+    POST values: {"bbox": {"lat1": number, "lat2": number, "lng1": number, "lng2": number}}
+
+    Returns
+    -------
+    Response
+        ACCEPTED is the expected response. Response body contains Celery taskId
+    """
+    try:
+        bbox = request.get_json()["bbox"]
+        lat1 = float(bbox.get("lat1"))
+        lng1 = float(bbox.get("lng1"))
+        lat2 = float(bbox.get("lat2"))
+        lng2 = float(bbox.get("lng2"))
+        scenario_options = request.get_json()["scenarioOptions"]
+    except ValueError:
+        return make_response(
+            "JSON values for bbox: lat1, lng1, lat2, lng2 must be valid floats", BAD_REQUEST
+        )
+    if any(coord is None for coord in [lat1, lng1, lat2, lng2]):
+        return make_response("JSON body parameters bbox: {lat1, lng1, lat2, lng2} mandatory", BAD_REQUEST)
+    if not valid_coordinates(lat1, lng1) or not valid_coordinates(lat2, lng2):
+        return make_response("lat & lng must fall in the range -90 < lat <= 90, -180 < lng <= 180", BAD_REQUEST)
+    if (lat1, lng1) == (lat2, lng2):
+        return make_response("lat1, lng1 must not equal lat2, lng2", BAD_REQUEST)
+
+    bbox_wkt = create_wkt_from_coords(lat1, lng1, lat2, lng2)
+    task = tasks.create_model_for_area(bbox_wkt, scenario_options)
+
+    return make_response(
+        jsonify({"taskId": task.id}),
+        ACCEPTED
+    )
+
+
+@flood_resilience_blueprint.route('/tasks/<task_id>/model/depth', methods=["GET"])
+@check_celery_alive
+def get_depth_at_point(task_id: str) -> Response:
+    """
[…]
