### 2024-11-22 85c423c
Update terriajs entrypoint to support Backend Hosts that use default ports.
