### 2024-11-24 9bc50bb
Combine base_data_to_db implementations
