### 2024-11-24 63e70da
Combine base_data_to_db implementations
