### 2024-11-24 faef85f
Ensure base data is initialised before MEDUSA model run, in case worker signals do not cause startup tasks to run.
