### 2024-11-24 09db887
Use stronger typing using generics on retry_function
