### 2024-11-24 b659df0
Include linting checks of extension modules in github actions
