+++ otakaro/tasks.py
@@ -5,14 +5,13 @@ Allows the frontend to send tasks and retrieve status later.
 import logging
 from typing import Dict, Union, Optional
 
-import sqlalchemy.exc
 from celery import signals
 from celery.worker.consumer import Consumer
 import geopandas as gpd
 
 from src.digitaltwin import retrieve_from_instructions
 from src.digitaltwin.utils import setup_logging
-from src.tasks import add_base_data_to_db, app, OnFailureStateTask, wkt_to_gdf
+from src.tasks import add_base_data_to_db, app, OnFailureStateTask, wkt_to_gdf  # pylint: disable=cyclic-import
 from otakaro import initialise_db_with_files
 from otakaro.environmental.water_quality import surface_water_sites
 from otakaro.pollution_model import run_medusa_2
