### 2024-11-24 6b07e12
250 modularise the architecture (#287)
* Split codebase by files into high-level core, floodresilience, otakaro

* Revert "Move pollution model initialisation to retrieve_static_boundaries.py"

This reverts commit ab33438f

* Split extension-specific code out of src.digitaltwin

* Fix linting

* Avoid use of inplace=True

* Separate web requests, celery tasks, and run_all scripts into separate modules for each project

* Combine base_data_to_db implementations

* Ensure base data is initialised before MEDUSA model run, in case worker signals do not cause startup tasks to run.

* Use stronger t
[…]
