+++ floodresilience/dynamic_boundary_conditions/rainfall/hirds_rainfall_data_from_db.py
@@ -1,0 +1,168 @@
+# -*- coding: utf-8 -*-
+"""Retrieve all rainfall data for sites within the catchment area from the database."""
+
+import logging
+from typing import Optional
+
+import geopandas as gpd
+import pandas as pd
+from sqlalchemy.engine import Engine
+from sqlalchemy.sql import text
+
+from floodresilience.dynamic_boundary_conditions.rainfall import hirds_rainfall_data_to_db
+
+log = logging.getLogger(__name__)
+
+
+def filter_for_duration(rain_data: pd.DataFrame, duration: str) -> pd.DataFrame:
+    """
+    Filter the HIRDS rainfall data for a requested duration.
+
+    Parameters
+    ----------
+    rain_data : pd.DataFrame
+        HIRDS rainfall data in Pandas DataFrame format.
+    duration : str
+        Storm duration. Valid options are: '10m', '20m', '30m', '1h', '2h', '6h', '12h', '24h', '48h', '72h',
+        '96h', '120h', or 'all'.
+
+    Returns
+    -------
+    pd.DataFrame
+        Filtered rainfall data for the requested duration.
+    """
+    if duration != "all":
+        # Filter the rain_data DataFrame to include only the relevant columns for the requested duration
+        rain_data = rain_data[["site_id", "category", "rcp", "time_period", "ari", "aep", duration]]
+    return rain_data
+
+
+def get_one_site_rainfall_data(
+        engine: Engine,
+        site_id: str,
+        rcp: Optional[float],
+        time_period: Optional[str],
+        ari: float,
+        duration: str,
+        idf: bool) -> pd.DataFrame:
+    """
+    Retrieve rainfall data from the database for the requested site based on the user-requested scenario.
+
+    Parameters
+    ----------
+    engine : Engine
+        The engine used to connect to the database.
+    site_id : str
+        HIRDS rainfall site ID.
+    rcp : Optional[float]
+        Representative Concentration Pathway (RCP) value. Valid options are 2.6, 4.5, 6.0, 8.5, or None
+        for historical data.
+    time_period : Optional[str]
+        Future time period. Valid options are "2031-2050", "2081-2100", or None for historical data.
+    ari : float
+        Average Recurrence Interval (ARI) value. Valid options are 1.58, 2, 5, 10, 20, 30, 40, 50, 60, 80, 100, or 250.
+    duration : str
+        Storm duration. Valid options are: '10m', '20m', '30m', '1h', '2h', '6h', '12h', '24h', '48h', '72h',
+        '96h', '120h', or 'all'.
+    idf : bool
+        Set to False for rainfall depth data, and True for rainfall intensity data.
+
+    Returns
+    -------
+    pd.DataFrame
+        Rainfall data for the requested site based on the user-requested scenario.
+
+    Raises
+    ------
+    ValueError
+        If rcp and time_period arguments are inconsistent.
+    """
+    # Get the relevant rainfall data table name from the idf parameter
+    rain_table_name = hirds_rainfall_data_to_db.db_rain_table_name(idf)
+    log.info(f"Retrieving the requested '{rain_table_name}' scenario data for site {site_id} from the database.")
+    # Check for inconsistent rcp and time_period arguments
+    if (rcp is None and time_period is not None) or (rcp is not None and time_period is None):
+        raise ValueError("Inconsistent arguments provided. "
+                         "For historical data, both 'rcp' and 'time_period' should be None. "
+                         "If 'rcp' is None, 'time_period' should also be None, and vice versa.")
+    elif rcp is not None and time_period is not None:
+        # Query for specific rcp and time_period
+        command_text = f"""
+        SELECT *
+        FROM {rain_table_name}
[…]
