### 2024-11-25 275c5bd
Create bounding box area calculation process
