### 2024-11-25 babe6b9
Add kowhaiwhai and red styling colours
