### 2024-11-25 9bb3b09
Add CCC road surface polygon dataset
