### 2024-11-25 8df9753
Remove nz_roads dataset from otakaro module
