### 2024-11-26 e3ec3ef
286 use ccc road surface dataset instead of stats nz road centreline dataset for medusa (#305)
* Add CCC road surface polygon dataset

* Remove nz_roads dataset from otakaro module

* Fix lint

---------

Co-authored-by: Luke Parkinson <luke.parkinson@canterbury.ac.nz>
