+++ otakaro/run_all.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 """This script runs each module in the Digital Twin using a Sample Polygon."""
-import pathlib
 
-from src.digitaltwin import retrieve_from_instructions
 from src.digitaltwin.utils import LogLevel
 from src.run_all import create_sample_polygon, main
 from otakaro import initialise_db_with_files
@@ -10,10 +8,6 @@ from otakaro.pollution_model import run_medusa_2
 from otakaro.environmental.water_quality import main_water_quality
 
 DEFAULT_MODULES_TO_PARAMETERS = {
-    retrieve_from_instructions: {
-        "log_level": LogLevel.INFO,
-        "instruction_json_path": pathlib.Path("otakaro/otakaro_data_instructions.json").as_posix()
-    },
     initialise_db_with_files: {
         "log_level": LogLevel.INFO,
     },
