### 2024-11-26 68ef123
Remove outline styling from MEDUSA polygons
