### 2024-11-26 dda4f59
Use db-name corrected gs workspace in terria (#315)
