### 2024-11-26 362c5f7
Add explanation for antecedent dry days which is (number of dry days between rain events) (#325)
Co-authored-by: Martin Nguyen <tmn52@uclive.ac.nz>
