### 2024-11-26 20ae0b8
Remove kowhaiwhai, set colours to avon green
