### 2024-11-26 a9824bd
draft codes to calculate the area of aoi
Use forked owslib to fix bounding box terria js incompatability

Implement basic area handler

Casey big PR
Format WPS bounding box area service

Fix OWSLib fork working in docker

Use parametised env variable for backend url in area process

issue #285 finalise data catalogue

issue #311 remove CREASSiteFishBarrier dataset

issue #316 split the items in the data catalogue into separate relevant tabs

update bounding box area calculation

fix lint

update area.py

issue #324 sort catalog group items alphabetically

use @staticmethod instead of self

not
[…]
