+++ otakaro/processes/__init__.py
@@ -1,0 +1,1 @@
+"""Defines PyWPS WebProcessingService processes for general use for the Ōtākaro digital twin."""
