+++ deploy/otakaro/cloudfront.py
@@ -1,0 +1,230 @@
+import json
+import shutil
+from pathlib import Path
+
+import pulumi
+import pulumi_aws as aws
+
+
+def create_records(domains, zone_name, dns_zone, cname_target):
+    for domain in domains:
+        record_name = domain.replace(f".{zone_name}", "")
+        aws.route53.Record(
+            f"{record_name}",
+            name=record_name,
+            zone_id=dns_zone.id,
+            type=aws.route53.RecordType.CNAME,
+            ttl=5,
+            records=[cname_target],
+        )
+
+
+class CloudFront:
+    """
+    Creates an AWS CloudFront distribution and associated Lambda, DNS Records, etc.
+
+    Parameters
+    ----------
+
+    aliases : str[list]
+        The list of domains that this CloudFront distribution supports.
+
+    origin_domain_name : str[list]
+        The URL that this CloudFront Distribution points to.
+
+    certificate : aws.acm.Certificate
+        An Amazon Issued HTTPS Certificate.
+
+    dns_zone : aws.route53.Zone
+        Zone to create DNS records in
+
+    zone_name : str
+        Required because retrieving the name from the dns zone is complicated.
+        This should just be the primary domain name (i.e otakaro.digitaltwins.nz as an alias
+        means this should be `digitaltwins.nz`)
+    """
+
+    def __init__(
+        self,
+        aliases: list[str],
+        origin_domain_name: str,
+        certificate: aws.acm.Certificate,
+        dns_zone: aws.route53.Zone,
+        zone_name: str,
+        basic_auth: tuple = None,
+    ):
+        # Cloudfront is us-east-1 only for the API
+        self.provider = aws.Provider("aws-east-1", region="us-east-1")
+        if basic_auth:
+            basic_auth_func = self._basic_auth(basic_auth)
+
+        cloudfront = aws.cloudfront.Distribution(
+            "cloudfront_dist",
+            aliases=[
+                "api.otakaro.digitaltwins.nz",
+                "otakaro.digitaltwins.nz",
+                "gs.otakaro.digitaltwins.nz",
+            ],
+            origins=[
+                aws.cloudfront.DistributionOriginArgs(
+                    domain_name=origin_domain_name,
+                    origin_id="otakaro-alb-internal",
+                    custom_origin_config=aws.cloudfront.DistributionOriginCustomOriginConfigArgs(
+                        http_port=80,
+                        https_port=443,
+                        origin_protocol_policy="https-only",
+                        origin_ssl_protocols=["TLSv1.1", "TLSv1.2"],
+                        origin_keepalive_timeout=5,
+                        origin_read_timeout=100,
+                    ),
+                    connection_timeout=10,
+                    connection_attempts=3,
+                )
+            ],
+            enabled=True,
+            default_cache_behavior=aws.cloudfront.DistributionDefaultCacheBehaviorArgs(
+                allowed_methods=[
+                    "GET",
+                    "HEAD",
+                    "PUT",
+                    "POST",
+                    "OPTIONS",
+                    "DELETE",
+                    "PATCH",
+                ],
+                cached_methods=["GET", "HEAD"],
+                target_origin_id="otakaro-alb-internal",
+                viewer_protocol_policy="redirect-to-https",
+                forwarded_values=aws.cloudfront.DistributionDefaultCacheBehaviorForwardedValuesArgs(
+                    cookies={"forward": "all"},
+                    query_string=True,
+                    headers=[
+                        "Host",
+                        "Origin",
+                        "Authorization",
+                    ],
+                ),
+                lambda_function_associations=(
+                    [
+                        aws.cloudfront.DistributionDefaultCacheBehaviorLambdaFunctionAssociationArgs(
[…]
