+++ deploy/otakaro/README.md
@@ -1,0 +1,170 @@
+# Otakaro AWS deployment 
+
+This documentation covers the deployment of the [Otakaro Digital Twin](https://otakaro.digitaltwins.nz/) to AWS.
+
+This directory contains a set of files that comprise a [Pulumi](https://www.pulumi.com/) deployment.
+
+All of the configuration is written in Python so as to reduce the overhead in bringing a developer up to speed with making changes to configuration where necessary.
+
+The modules in this repository can best be summed up as such:
+
+ - `__main__.py`: This is where the bulk of the config happens, pulling in modules as required and contains a lot of the configuration. If you're not deploying new services, this file is where you'll spend most of your time.
+ - `efs.py`: Creates [Amazon EFS](https://aws.amazon.com/efs/) file shares. These get mounted to services as NFS shares.
+ - `alb.py`: Creates an [Amazon Application Load Balancer](https://docs.aws.amazon.com/elasticloadbalancing/latest/application/application-load-balancers.html) for routing requests to ECS services.
+ - `certs.py`: Automates the creation of HTTPS certificates with Amazon as the CA.
+ - `postgres.py`: Creates [Amazon RDS](https://aws.amazon.com/rds/postgresql/) instances, specifically PostgreSQL.
+ - `cloudfront.py`: Sets up [Amazon CloudFront](https://aws.amazon.com/cloudfront/) Distributions. This can also set up Basic Authentication as a [Lambda@Edge](https://aws.amazon.com/lambda/edge/) function.
+ - `fargate.py`: Set up everything required to run an instance of a container in [ECS Fargate](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/AWS_Fargate.html). This sets up a container repository, a lambda function to update the image tag and a 
+ - `backend.py`: Specifically sets up all of the extra resources that the Otakaro Digital Twin backend requires. Stands up 2 ECS Services (Celery Worker + Flask API Server) and a Redis instance to handle Message Queueing and Brokering.
+
+## Getting Started
+
+To get started it's assumed you have access to a Linux command line in some way shape or form (i.e WSL2 or Linux. If you're using a Mac that's basically Unix.)
+
+### Installing Pulumi
+The instructions can be [found here](https://www.pulumi.com/docs/iac/download-install/).
+
+Once that's done you should be able to do something like this in a terminal. (The version will probably be newer for you)
+```
+ $ pulumi version
+v3.137.0
+```
+
+### How Pulumi Manages State
+Pulumi uses a state file to keep state, and you "log into" a Pulumi backend. 
+
+For this codebase assume we're using an S3 bucket as our backend, because we're already in an AWS environment anyway.
+
+If you want more information, it can be [found here](https://www.pulumi.com/docs/iac/concepts/state-and-backends/).
+
+### Setting up Pulumi 
+
+For this we're going to assume we're using the current Otakaro environment, but it's possible to remove 
+
+For the Otakaro Digital Twin environment specifically, you're going to need a password. This has been provided to Luke Parkinson, so get it from him. 
+
+Firstly, you need an active AWS profile in your environment, and you need to set it in your environment. 
+
+This can be done with 
+```
+export AWS_PROFILE=[insert-profile-name-here]
+```
+
+Setting up an AWS Profile is beyond the scope of this documentation.
+
+Next, you need to "log into" the pulumi state. For our purposes, the Otakaro bucket will be used as an example:
+```
+ $ pulumi login s3://otakaro-dt-pulumi-state
+Logged in to vettel as alistairmcintyre (s3://otakaro-dt-pulumi-state)
+```
+
[…]

[…]
