### 2024-11-27 477bc76
Fix styling issues found during testing
