### 2024-11-27 7526b93
Add extra missing text to pulumi README
