### 2024-11-27 bbc0294
terriajs add geometry functionalities to mark points and calculate areas
