### 2024-11-27 69f2fa7
remove wps process for calculating bounding box area
