### 2024-11-27 ee6a1bd
if geometry is unknown display the output details in lightgreen
