### 2024-11-28 c9a7d18
update GeometryProcessService output details for unknown geom tryp
