+++ otakaro/pollution_model/run_medusa_2.py
@@ -1,923 +1,0 @@
-"""
-This script takes in appropriate datasets as input.
-Then runs the MEDUSA2.0 model to calculate TSS (total suspended solids), TCu (total copper), DCu  (dissolved copper),
-TZn (total zinc), and DZn (dissolved zinc).
-
-DOI of the model paper: https://doi.org/10.3390/w12040969
-"""
-
-import logging
-import math
-from enum import StrEnum
-from typing import NamedTuple, Dict, Union, Optional
-from xml.sax import saxutils
-
-import geopandas as gpd
-import pandas as pd
-from sqlalchemy.engine import Engine
-from sqlalchemy.sql import text
-from tqdm import tqdm
-
-from src import geoserver
-from src.config import EnvVariable
-from src.digitaltwin import setup_environment
-from src.digitaltwin.tables import create_table, check_table_exists, execute_query
-from src.digitaltwin.utils import get_catchment_area, LogLevel, setup_logging
-from otakaro.pollution_model.pollution_tables import Medusa2ModelOutputBuildings, Medusa2ModelOutputRoads
-from otakaro.pollution_model.pollution_tables import MedusaScenarios
-
-log = logging.getLogger(__name__)
-
-
-# Enum strings are assigned as they are described in the original paper
-class SurfaceType(StrEnum):
-    """
-    StrEnum to represent the surface types that a feature could have.
-
-    Attributes
-    ----------
-    COLORSTEEL : str
-        Colorsteel surface
-    GALVANISED : str
-        Galvanised surface.
-    METAL_OTHER : str
-        Metal Other surface.
-    METAL_TILE : str
-        Metal Tile surface.
-    NON_METAL : str
-        Non-Metal surface.
-    ZINCALUME : str
-        Zincalume surface.
-    ASPHALT_ROAD: str
-        Asphalt Road.
-    CAR_PARK : str
-        Car park.
-    """
-
-    COLORSTEEL = "Colorsteel"
-    GALVANISED = "Galvanised"
-    METAL_OTHER = "MetalOther"
-    METAL_TILE = "MetalTile"
-    NON_METAL = "NonMetal"
-    ZINCALUME = "Zincalume"
-    ASPHALT_ROAD = "AsphaltRoad"
-    CAR_PARK = "CarPark"  # CarParks are classified the same as roads
-
-
-ROOF_SURFACE_TYPES = {SurfaceType.COLORSTEEL, SurfaceType.GALVANISED, SurfaceType.METAL_OTHER, SurfaceType.METAL_TILE,
-                      SurfaceType.NON_METAL, SurfaceType.ZINCALUME}
-ROAD_SURFACE_TYPES = {SurfaceType.ASPHALT_ROAD, SurfaceType.CAR_PARK}
-
-
-class MedusaRainfallEvent(NamedTuple):
-    """
-    Rainfall event parameters for MEDUSA 2.0 model.
-
-    Attributes
-    ----------
-    antecedent_dry_days: float
-        Length of antecedent dry period (days).
-    average_rain_intensity: float
-        Average rainfall intensity of the event (mm/h).
-    event_duration: float
-        Duration of the rainfall event (h).
-    rainfall_ph: float
-        The acidity level of the rainfall.
-    """
-
-    antecedent_dry_days: float
-    average_rain_intensity: float
-    event_duration: float
-    rainfall_ph: float
-
-    def as_dict(self) -> Dict[str, float]:
-        """
-        Convert the MedusaRainfallEvent parameters to a dictionary.
-
-        Returns
-        -------
-        Dict[str, float]
-            Returns the MedusaRainfallEvent parameters as a dictionary.
-        """
[…]
