+++ otakaro/environmental/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Package to handle data and operations specific to environmental data,
-such as water quality.
-"""
