### 2024-11-28 3e17d6b
Full Pulumi Config (#327)
* Full Pulumi Config

* Add extra missing text to pulumi README

---------

Co-authored-by: Alistair McIntyre <alistair@openplan.co.nz>
