+++ deploy/otakaro/fargate.py
@@ -1,0 +1,341 @@
+import json
+
+import pulumi
+import pulumi_aws as aws
+import pulumi_awsx as awsx
+
+
+class FargateApplication:
+    """
+    Creates a default Fargate Application on Amazon ECS. Also automatically instantiates an ECR
+    Image Repository, a Load Balancer Target Group (if a port is supplied) and a Lambda Function to
+    automatically refresh the running tasks images.
+
+    NOTE: Some manual intervention is required to enable the trigger for the lambda. At this stage
+    there's no good documentation on how to do this with Pulumi and it's just easier to connect it
+    in the AWS Console.
+
+    Parameters
+    ----------
+
+    service_name : str
+        The name of the ECS Fargate Service you wish to run
+
+    cluster : aws.ecs.Cluster
+        The Cluster that this service will run within.
+
+    vpc : aws.ec2.Vpc
+        The VPC that your Cluster is connected to.
+
+    service_port : int
+        The port that your application listens on for HTTP traffic. Defaults to None.
+        If not specified, no Target Group will be created for a load balancer.
+
+    image_tag : str
+        The tag of the container image within the supplied or created repository.
+
+    repository : aws.ecr.Repository
+        If this is set, no new ECR Repository will be created, and this one will be used instead.
+
+    auto_image_push : bool
+        If set to True, an AWS Lambda function is created that will automatically update the image
+        tag to a new tag on image push.
+
+    entrypoint : list[str]
+        The entrypoint command for the container if it needs to be overridden.
+
+    environment : list[dict]
+        Environment variables for the container. Secrets should not be set here.
+
+    replica_count : int
+        How many "tasks" (i.e replicas of your container) you want to run
+
+    cpu : int
+        How much CPU should this task get? 1000 ~= 1 vCPU.
+
+    memory : int
+        How much Memory should this task get? 1024 ~= 1GB Mem
+
+    volumes : list[aws.ecs.TaskDefinitionVolumeArgs]
+        The list of volumes that this task has access to
+
+    mount_points : list[awsx.ecs.TaskDefinitionMountPointArgs]
+        The list of mount points for the volumes specified above.
+
+    secrets : list[awsx.ecs.TaskDefinitionSecretArgs]
+        The list of secrets to mount as Environment Variables.
+    """
+
+    def __init__(
+        self,
+        service_name: str,
+        cluster: aws.ecs.Cluster,
+        vpc: awsx.ec2.Vpc,
+        service_port: int = None,
+        image_tag: str = "latest",
+        repository: aws.ecr.Repository = None,
+        auto_image_push: bool = True,
+        entrypoint: list[str] = None,
+        environment: list[dict] = [],
+        replica_count: int = 2,
+        cpu: int = 128,
+        memory: int = 512,
+        volumes: list[aws.ecs.TaskDefinitionVolumeArgs] = [],
+        mount_points: list[awsx.ecs.TaskDefinitionMountPointArgs] = [],
+        secrets: [awsx.ecs.TaskDefinitionSecretArgs] = [],
+    ):
+        repo = repository
+        if not repo:
+            repo_name = f"{service_name}-ecr"
+            repo = aws.ecr.Repository(
+                repo_name,
+                name=repo_name,
+                image_tag_mutability="MUTABLE",
+                image_scanning_configuration={
+                    "scan_on_push": True,
+                },
+            )
+
+        security_group = aws.ec2.SecurityGroup(
+            f"{service_name}-sg",
+            vpc_id=vpc.vpc_id,
+            egress=[
+                aws.ec2.SecurityGroupEgressArgs(
+                    from_port=0,
+                    to_port=0,
+                    protocol="-1",
+                    cidr_blocks=["0.0.0.0/0"],
+                    ipv6_cidr_blocks=["::/0"],
[…]
