+++ deploy/otakaro/Pulumi.yaml
@@ -1,0 +1,11 @@
+name: otakaro
+description: A minimal AWS Python Pulumi program
+runtime:
+  name: python
+  options:
+    toolchain: pip
+    virtualenv: venv
+config:
+  pulumi:tags:
+    value:
+      pulumi:template: aws-python

[…]
