### 2024-11-28 3fb2ff3
Change email to shared mailbox
