+++ src/pywps.cfg
@@ -14,7 +14,7 @@ contact_city=Christchurch
 contact_stateorprovince=None
 contact_postalcode=8140
 contact_country=New Zealand
-contact_email=luke.parkinson@canterbury.ac.nz
+contact_email=gri-enquiries+otakaro@canterbury.ac.nz
 contact_instructions=Contact via Email or GitHub
 contact_role=pointOfContact
 
