### 2024-11-29 0f9513a
Fix styling inconsistencies
