### 2024-12-11 53f96a9
issue #267 update the river module to utilize the `arcgis_rest_api.py` script
