+++ floodresilience/dynamic_boundary_conditions/river/river_data_from_niwa.py
@@ -1,241 +1,40 @@
 # -*- coding: utf-8 -*-
 """Fetch REC data in New Zealand from NIWA using the ArcGIS REST API."""
 
-import asyncio
 import logging
-from typing import List, Dict, Union, NamedTuple
 
-import aiohttp
 import geopandas as gpd
-import pandas as pd
 import requests
-from shapely.geometry import LineString
-from sqlalchemy.engine import Engine
 
-from src.digitaltwin.utils import get_nz_boundary
+from src.digitaltwin.arcgis_rest_api import fetch_arcgis_rest_api_data
 
 log = logging.getLogger(__name__)
 
-# URL for retrieving REC data from NIWA using the ArcGIS REST API
-REC_API_URL = "https://gis.niwa.co.nz/server/rest/services/HYDRO/Flood_Statistics_Henderson_Collins_V2/MapServer/2"
 
-
-class RecordCounts(NamedTuple):
-    """
-    Represents the record counts of the REC feature layer.
-
-    Attributes
-    ----------
-    max_record_count : int
-        The maximum number of records that will be returned per query.
-    total_record_count : int
-        The total number of records available in the feature layer.
-    """  # pylint: disable=duplicate-code
-
-    max_record_count: int
-    total_record_count: int
-
-
-def get_feature_layer_record_counts(url: str = REC_API_URL) -> RecordCounts:
-    """
-    Retrieve the maximum and total record counts from the REC feature layer.
-
-    Parameters
-    ----------
-    url : str = REC_API_URL
-        The URL of the REC feature layer. Defaults to `REC_API_URL`.
-
-    Returns
-    -------
-    RecordCounts
-        A named tuple containing the maximum and total record counts of the REC feature layer.
-
-    Raises
-    ------
-    RuntimeError
-        If there is an issue with retrieving the record counts from the REC feature layer.
-    """  # pylint: disable=duplicate-code
-    # Set up parameters for the initial request to get the maximum record count
-    params = {"f": "json"}
-    response = requests.get(url=url, params=params)
-    # Extract the maximum record count from the response
-    max_record_count = response.json()["maxRecordCount"]
-    # Set up parameters for the second request to get the total record count
-    params["where"] = "1=1"
-    params["returnCountOnly"] = True
-    response = requests.get(url=f"{url}/query", params=params)
-    try:
-        # Extract the total record count from the response
-        total_record_count = response.json()["count"]
-    except KeyError as e:
-        # Raise a RuntimeError to indicate the API failure
-        raise RuntimeError("Failed to fetch rec data feature layer record counts.") from e
-    # Returns the maximum and total record counts of the REC feature layer
-    return RecordCounts(max_record_count, total_record_count)
-
-
-def gen_rec_query_param_list(
-        engine: Engine,
-        max_record_count: int,
-        total_record_count: int) -> List[Dict[str, Union[str, int]]]:
-    """
-    Generate a list of API query parameters used to retrieve REC data in New Zealand.
-
-    Parameters
-    ----------
-    engine : Engine
-        The engine used to connect to the database.
-    max_record_count : int
-        The maximum number of records that will be returned per query.
-    total_record_count : int
-        The total number of records available in the feature layer.
-
-    Returns
-    -------
-    List[Dict[str, Union[str, int]]]
-        A list of API query parameters used to retrieve REC data in New Zealand.
-    """
-    # Get the New Zealand boundary geometry in the specified CRS
-    nz_boundary = get_nz_boundary(engine, to_crs=2193)
-    # Extract the bounding box coordinates from the New Zealand boundary
-    x_min, y_min, x_max, y_max = nz_boundary.total_bounds
[…]
