### 2024-12-20 ed9c2c9
Add copyright notice to each non-trivial source file
