### 2024-12-20 GeospatialResearch/Digital-Twins#279 @LukeParky
Hi @pritamd47, thank you for suggesting this. A fix is available in the PR #341. Could you please check that it matches your expectations for this issue?
