### 2024-12-20 GeospatialResearch/Digital-Twins#279 @LukeParky
Note: This issue is related to the JOSS submission openjournals/joss-reviews#7433
