### 2024-12-20 #7433 @LukeParky
Hi @pritamd47, for your review items below, could you please expand on these?
Have you encountered errors, or is there another barrier to installation that impacts review?

> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
