### 2024-12-20 867d902
Add notes about support and contributing to the README.md
