### 2024-12-20 f486498
Add notes about support and contributing to the README.md
