### 2024-12-25 #7433 @pritamd47
> Hi @pritamd47, for your review items below, could you please expand on these? Have you encountered errors, or is there another barrier to installation that impacts review?
>
> > ### Functionality
> >
> > * [ ]   **Installation:** Does installation proceed as outlined in the documentation?
> > * [ ]   **Functionality:** Have the functional claims of the software been confirmed?

Hi @LukeParky, I was following older version of the documentation on installing the software due to which I couldn't install it. earlier. I see that you have updated the documentation since. I tried installing the software again and now I think I am getting an error because the "POSTGRES_PASSWORD" field in my .env file is empty. How do I get this password?
