### 2024-12-28 #7433 @LukeParky
> I tried installing the software again and now I think I am getting an error because the "POSTGRES_PASSWORD" field in my .env file is empty. How do I get this password?

This is a password that you configure yourself. Anything will be fine. Sorry for my delay in replying, I am not working for the next few weeks but I am intending to answer your questions as they come in. I have created an issue #342 to track this comment, and put up a PR #342 that I think resolves it, if you would like to look at this and let me know if it clarifies things.
