### 2024-12-31 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi @LukeParky, thanks for creating the issue. I was able to resolve this issue. However, I failed to run it on my computer which is mac (m1 chip) and got an error of mismatch in architecture when I tried to run docker. I was able to run it on another machine using WSL (linux emulation for windows). However, I was still getting some errors. Please take a look at the attached screenshots.

As suggested by documentation, when I visit 8080 port, I get the following error -
![image (1)](https://github.com/user-attachments/assets/6b815343-ee0f-41bd-9ac7-1c63982e2f5b)

From the docker GUI I opened another link to a port that seemed to load the tool but there I got the following error -
![image](https://github.com/user-attachments/assets/378c305f-50c7-4761-8c9e-be3073e089fe)

Could you help me with these errors while installing the software?
