### 2025-01-06 #7433 @pritamd47
Hi @cheginit. I am trying to install the software, where I am running into errors. @LukeParky has created the following issue [342](https://github.com/GeospatialResearch/Digital-Twins/issues/342) where we are working to resolve the issue before finalizing the review.
