### 2025-01-15 GeospatialResearch/Digital-Twins#344 @changliao1025
Task failed in the backend. Please try again.
When I tried to run the "Generating model", it always failed with the permission issue. See below log from the terminal:

Arguments: 'src.tasks.process_dem', 'exc': "PermissionError(13, 'Permission denied')", 'traceback': 'Traceback (most recent call last):\n  File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 451, in trace_task\n    R = retval = fun(*args, **kwargs)\n                 ^^^^^^^^^^^^^^^^^^^^\n  File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 734, in __protected_call__\n    return self.run(*args, **kwargs)\n           ^^^^^^^^^^^^^^^^^^^^^^^^^\n  File "/app/src/tasks.py", line 126, in process_dem\n    process_hydro_dem.main(selected_polygon, **parameters)\n  File "/app/src/flood_model/process_hydro_dem.py", line 106, in main\n    process_dem(selected_polygon_gdf)\n  File "/app/src/flood_model/process_hydro_dem.py", line 62, in process_dem\n    newzealidar.process.main(selected_polygon_gdf)\n  File "/venv/lib/python3.11/site-packages/newzealidar/process.py", line 498, in main\n    utils.save_gpkg(lidar_extent, "lidar_extent")\n  File "/venv/lib/python3.11/site-packages/newzealidar/utils.py", line 994, in save_gpkg\n    pathlib.Path(gpkg_path).mkdir(parents=True, exist_ok=True)\n  File "/venv/lib/python3.11/pathlib.py", line 1116, in mkdir\n    os.mkdir(self, mode)\nPermissionError: [Errno 13] Permission denied: \'/stored_data/gpkg\'\n', 'args': "['POLYGON ((172.71629837906798 -43.40100735095039, 172.71629837906798 -43.38586581746695, 172.67209098399104 -43.38586581746695, 172.67209098399104 -43.40100735095039, 172.71629837906798 -43.40100735095039))']", 'kwargs': '{}', 'description': 'raised unexpected', 'internal': False}
