### 2025-01-15 #7433 @changliao1025
> @changliao1025 a friendly reminder for this review.

Submitted an issue: https://github.com/GeospatialResearch/Digital-Twins/issues/344
