### 2025-01-19 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47, thank you for making aware of the issue about mac, this may be able to be resolved by building the image manually but I do not have access to a mac to test.
I will provide an update to the documentation for the web port.

As for the error message you are receiving, could you please provide your logs with `docker compose logs -f backend celery_worker`?
Do you still receive the error message?
