### 2025-01-19 GeospatialResearch/Digital-Twins#342 @LukeParky
You may run into the issue that the other reviewer is encountering, https://github.com/GeospatialResearch/Digital-Twins/issues/344, working on a fix
