### 2025-01-19 GeospatialResearch/Digital-Twins#344 @LukeParky
Hi @changliao1025 thank you, it looks like a regression has caused this. I am investigating now.
