### 2025-01-20 df0adb7
Increase explanation of .env variab;es (#343)
