### 2025-01-20 26ddd10
Add copyright notice to each non-trivial source file (#341)
* Add copyright notice to each non-trivial source file

* Update copyright to 2025
