### 2025-01-20 851a980
Lock conda-pack because newer versions break builds (#345)
* Lock conda-pack because newer versions break builds

* Bump docker tag version numbers
