### 2025-01-20 GeospatialResearch/Digital-Twins#344 @LukeParky
Hi @changliao1025 I believe the issue has been resolved. You will have to pull/build a new image, and I reccomend starting with fresh volumes.

Eg.
`docker compose pull && docker compose down -v && docker compose up -d && docker compose logs -f backend celery_worker`
