### 2025-01-20 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47 I recommend git pulling from the newest changes, and then pulling/building a new docker image.
I am not aware of any further issues that should impact your review but please reach out if you are having trouble.
