### 2025-01-20 #7433 @LukeParky
I am now back and available for work on this.

I believe I have addressed the functional setup concerns that have occurred for the reviewers, but I am awaiting their confirmation. The issue was likely caused by a recent external dependency update.

Next I will address the paper comments, which should not take too long.
