### 2025-01-22 9ddc874
HOTFIX: use protocol/origin from window (#346)
* Add notes about support and contributing to the README.md

* Use protocol and origin from window

* Bunp www docker tag to 1.3.1
