+++ docker-compose.yml
@@ -116,7 +116,7 @@ services:
 
   www:
     # Webserver for the website interface
-    image: lparkinson/www-flood-resilience-dt:1.3
+    image: lparkinson/www-flood-resilience-dt:1.3.1
     build:
       context: .
       dockerfile: visualisation/Dockerfile
