### 2025-01-23 e86d3cf
Remove otakaro module
