+++ otakaro/environmental/water_quality/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Package to handle data and operations specific to water quality."""
