+++ otakaro/environmental/water_quality/surface_water_sites.py
@@ -1,103 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Fetch surface water site data from ECAN using the ArcGIS REST API, store it in the database, and
-retrieve it for the requested area of interest.
-"""
-
-import logging
-
-import geopandas as gpd
-from sqlalchemy.engine import Engine
-from sqlalchemy.sql import text
-
-from src.digitaltwin import setup_environment
-from src.digitaltwin.arcgis_rest_api import fetch_arcgis_rest_api_data
-from src.digitaltwin.tables import check_table_exists
-
-log = logging.getLogger(__name__)
-
-# The URL for retrieving surface water site data from ECAN
-SURFACE_URL = "https://gis.ecan.govt.nz/arcgis/rest/services/Public/WaterQualityandMonitoring/MapServer/0"
-
-
-def refresh_surface_water_sites() -> None:
-    """
-    Fetch surface water site data from ECAN and store it in the database.
-    Needs to be run periodically so that the surface water site data is up to date.
-    """
-    # Connect to the database
-    engine = setup_environment.get_database()
-    # Define the table name for storing the surface water site data
-    table_name = "surface_water_sites"
-    # The URL for retrieving surface water site data from ECAN
-    # Fetch surface water site data for the area of interest using the ArcGIS REST API
-    surface_sites = fetch_arcgis_rest_api_data(SURFACE_URL)
-    # Store the surface water site data in the relevant table in the database
-    log.info(f"Adding '{table_name}' to the database.")
-    surface_sites.to_postgis(table_name, engine, index=False, if_exists="replace")
-    log.info(f"Successfully added '{table_name}' to the database.")
-
-
-def store_surface_water_sites_to_db(
-        engine: Engine,
-        area_of_interest: gpd.GeoDataFrame = None,
-        output_sr: int = None) -> None:
-    """
-    Fetch surface water site data from ECAN using the ArcGIS REST API and store it in the database.
-
-    Parameters
-    ----------
-    engine : Engine
-        The engine used to connect to the database.
-    area_of_interest : gpd.GeoDataFrame = None
-        A GeoDataFrame representing the area of interest for data retrieval. If not provided, all data will be fetched.
-    output_sr : int = None
-        The EPSG code of the spatial reference system in which the requested data should be returned if no area of
-        interest is provided.
-    """
-    # Define the table name for storing the surface water site data
-    table_name = "surface_water_sites"
-    # Check if the table already exists in the database
-    if check_table_exists(engine, table_name):
-        log.info(f"'{table_name}' already exists in the database.")
-    else:
-        # Fetch surface water site data for the area of interest using the ArcGIS REST API
-        surface_sites = fetch_arcgis_rest_api_data(SURFACE_URL, area_of_interest, output_sr)
-        # Store the surface water site data in the relevant table in the database
-        log.info(f"Adding '{table_name}' to the database.")
-        surface_sites.to_postgis(table_name, engine, index=False, if_exists="replace")
-        log.info(f"Successfully added '{table_name}' to the database.")
-
-
-def get_surface_water_sites_from_db(engine: Engine, catchment_area: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
-    """
-    Retrieve surface water site data from the database for the specified catchment area.
-
-    Parameters
-    ----------
-    engine : Engine
-        The engine used to connect to the database.
-    catchment_area : gpd.GeoDataFrame
-        A GeoDataFrame representing the catchment area.
-
-    Returns
-    -------
-    gpd.GeoDataFrame
-        A GeoDataFrame containing the retrieved surface water site data for the requested catchment area.
-    """
[…]
