+++ src/tasks.py
@@ -100,5 +100,4 @@ def wkt_to_gdf(wkt: str) -> gpd.GeoDataFrame:
 
 
 # These must be imported after app to remove a circular dependency
-import otakaro.tasks  # pylint: disable=wrong-import-position,unused-import # noqa: E402, F401
 import floodresilience.tasks  # pylint: disable=wrong-import-position,unused-import # noqa: E402, F401

[…]
