### 2025-01-23 71f2ae3
First draft of wps format for FReDT
