### 2025-01-24 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi Luke. I pulled the latest changes as you instructed and started ran the `docker compose up -d` command. The container seems to be running.

![Image](https://github.com/user-attachments/assets/93c3ddec-6549-4eb8-8c6c-9e568da740d9)

However, localhost:8080 throws a connection refused error. The celery_logs are attached below.
```
backend_digital_twin        | [2025-01-24 07:18:49 +0000] [1] [INFO] Starting gunicorn 20.1.0
backend_digital_twin        | [2025-01-24 07:18:49 +0000] [1] [INFO] Listening at: http://0.0.0.0:5000 (1)
backend_digital_twin        | [2025-01-24 07:18:49 +0000] [1] [INFO] Using worker: sync
backend_digital_twin        | [2025-01-24 07:18:49 +0000] [13] [INFO] Booting worker with pid: 13
backend_digital_twin        | /venv/lib/python3.11/site-packages/geopandas/_compat.py:123: UserWarning: The Shapely GEOS version (3.11.2-CAPI-1.17.2) is incompatible with the GEOS version PyGEOS was compiled with (3.10.4-CAPI-1.16.2). Conversions between both will be slow.
backend_digital_twin        |   warnings.warn(
backend_digital_twin        | /app/src/tasks.py:25: UserWarning: Shapely 2.0 is installed, but because PyGEOS is also installed, GeoPandas will still use PyGEOS by default for now. To force to use and test Shapely 2.0, you have to set the environment variable USE_PYGEOS=0. You can do this before starting the Python process, or in your code before importing geopandas:
backend_digital_twin        |
backend_digital_twin        | import os
backend_digital_twin        | os.environ['USE_PYGEOS'] = '0'
backend_digital_twin        | import geopandas
backend_digital_twin        |
backend_digital_twin        | In a future release, GeoPandas will switch to using Shapely by default. If you are using PyGEOS directly (calling PyGEOS functions on geometries from GeoPandas), this will then stop working and you are encouraged to migrate from PyGEOS to Shapely 2.0 (https://shapely.readthedocs.io/en/latest/migration_pygeos.html).
backend_digital_twin        |   import geopandas as gpd
celery_worker_digital_twin  | /venv/lib/python3.11/site-packages/geopandas/_compat.py:123: UserWarning: The Shapely GEOS version (3.11.2-CAPI-1.17.2) is incompatible with the GEOS version PyGEOS was compiled with (3.10.4-CAPI-1.16.2). Conversions between both will be slow.
celery_worker_digital_twin  |   warnings.warn(
celery_worker_digital_twin  | /app/src/tasks.py:25: UserWarning: Shapely 2.0 is installed, but because PyGEOS is also installed, GeoPandas will still use PyGEOS by default for now. To force to use and test Shapely 2.0, you have to set the environment variable USE_PYGEOS=0. You can do this before starting the Python process, or in your code before importing geopandas:
celery_worker_digital_twin  |
celery_worker_digital_twin  | import os
celery_worker_digital_twin  | os.environ['USE_PYGEOS'] = '0'
celery_worker_digital_twin  | import geopandas
celery_worker_digital_twin  |
celery_worker_digital_twin  | In a future release, GeoPandas will switch to using Shapely by default. If you are using PyGEOS directly (calling PyGEOS functions on geometries from GeoPandas), this will then stop working and you are encouraged to migrate from PyGEOS to Shapely 2.0 (https://shapely.readthedocs.io/en/latest/migration_pygeos.html).
celery_worker_digital_twin  |   import geopandas as gpd
celery_worker_digital_twin  |
celery_worker_digital_twin  |  -------------- celery@719d983c4d65 v5.2.7 (dawn-chorus)
celery_worker_digital_twin  | --- ***** -----
celery_worker_digital_twin  | -- ******* ---- Linux-5.15.153.1-microsoft-standard-WSL2-x86_64-with-glibc2.35 2025-01-24 07:18:53
celery_worker_digital_twin  | - *** --- * ---
celery_worker_digital_twin  | - ** ---------- [config]
celery_worker_digital_twin  | - ** ---------- .> app:         tasks:0x7f5773591dd0
celery_worker_digital_twin  | - ** ---------- .> transport:   redis://message_broker:6379/0
celery_worker_digital_twin  | - ** ---------- .> results:     redis://message_broker:6379/0
celery_worker_digital_twin  | - *** --- * --- .> concurrency: 16 (thread)
celery_worker_digital_twin  | -- ******* ---- .> task events: OFF (enable -E to monitor tasks in this worker)
celery_worker_digital_twin  | --- ***** -----
celery_worker_digital_twin  |  -------------- [queues]
celery_worker_digital_twin  |                 .> celery           exchange=celery(direct) key=celery
celery_worker_digital_twin  |
celery_worker_digital_twin  |
celery_worker_digital_twin  | [tasks]
celery_worker_digital_twin  |   . src.tasks.add_base_data_to_db
celery_worker_digital_twin  |   . src.tasks.generate_rainfall_inputs
celery_worker_digital_twin  |   . src.tasks.generate_river_inputs
celery_worker_digital_twin  |   . src.tasks.generate_tide_inputs
celery_worker_digital_twin  |   . src.tasks.get_depth_by_time_at_point
celery_worker_digital_twin  |   . src.tasks.get_model_extents_bbox
celery_worker_digital_twin  |   . src.tasks.get_model_output_filepath_from_model_id
celery_worker_digital_twin  |   . src.tasks.get_valid_parameters_based_on_confidence_level
celery_worker_digital_twin  |   . src.tasks.process_dem
celery_worker_digital_twin  |   . src.tasks.refresh_lidar_datasets
celery_worker_digital_twin  |   . src.tasks.run_flood_model
celery_worker_digital_twin  |   . src.tasks.validate_slr_parameters
celery_worker_digital_twin  |
celery_worker_digital_twin  | (PDAL Error) Can't load library /venv/lib/libpdal_plugin_reader_draco.so: Failed to load "/venv/lib/libpdal_plugin_reader_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_writer_draco.so: Failed to load "/venv/lib/libpdal_plugin_writer_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_reader_draco.so: Failed to load "/venv/lib/libpdal_plugin_reader_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_writer_draco.so: Failed to load "/venv/lib/libpdal_plugin_writer_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory[2025-01-24 07:18:53,428: INFO/MainProcess] Connected to redis://message_broker:6379/0
celery_worker_digital_twin  | [2025-01-24 07:18:53,430: INFO/MainProcess] mingle: searching for neighbors
celery_worker_digital_twin  | [2025-01-24 07:18:54,438: INFO/MainProcess] mingle: all alone
celery_worker_digital_twin  | [2025-01-24 07:18:54,460: INFO/MainProcess] celery@719d983c4d65 ready.
backend_digital_twin        | (PDAL Error) Can't load library /venv/lib/libpdal_plugin_reader_draco.so: Failed to load "/venv/lib/libpdal_plugin_reader_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_writer_draco.so: Failed to load "/venv/lib/libpdal_plugin_writer_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_reader_draco.so: Failed to load "/venv/lib/libpdal_plugin_reader_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory(PDAL Error) Can't load library /venv/lib/libpdal_plugin_writer_draco.so: Failed to load "/venv/lib/libpdal_plugin_writer_draco.so": libdraco.so.8: cannot open shared object file: No such file or directory[2025-01-24 07:22:10 +0000] [1] [CRITICAL] WORKER TIMEOUT (pid:13)
backend_digital_twin        | [2025-01-24 07:22:10 +0000] [13] [INFO] Worker exiting (pid: 13)
backend_digital_twin        | [2025-01-24 07:22:10 +0000] [160] [INFO] Booting worker with pid: 160
backend_digital_twin        | /venv/lib/python3.11/site-packages/geopandas/_compat.py:123: UserWarning: The Shapely GEOS version (3.11.2-CAPI-1.17.2) is incompatible with the GEOS version PyGEOS was compiled with (3.10.4-CAPI-1.16.2). Conversions between both will be slow.
backend_digital_twin        |   warnings.warn(
backend_digital_twin        | /app/src/tasks.py:25: UserWarning: Shapely 2.0 is installed, but because PyGEOS is also installed, GeoPandas will still use PyGEOS by default for now. To force to use and test Shapely 2.0, you have to set the environment variable USE_PYGEOS=0. You can do this before starting the Python process, or in your code before importing geopandas:
backend_digital_twin        |
backend_digital_twin        | import os
backend_digital_twin        | os.environ['USE_PYGEOS'] = '0'
backend_digital_twin        | import geopandas
backend_digital_twin        |
backend_digital_twin        | In a future release, GeoPandas will switch to using Shapely by default. If you are using PyGEOS directly (calling PyGEOS functions on geometries from GeoPandas), this will then stop working and you are encouraged to migrate from PyGEOS to Shapely 2.0 (https://shapely.readthedocs.io/en/latest/migration_pygeos.html).
backend_digital_twin        |   import geopandas as gpd

```
