### 2025-01-27 362cce9
Add inputs for main fredt process
