### 2025-01-28 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47, if localhost:8080 then perhaps your port does not match. Could you please provide the output of `docker compose ps`? It should tell you what port to use.

If that is not the issue, could you please take a screenshot of the connection refused error?
It may be good for me to provide some one-on-one support to make sure this review can be completed, and see if there are ways we can improve the setup instructions, which time-zone are you in?
