### 2025-02-03 23039f0
Add ability to start model from terria
