### 2025-02-03 7e85b4b
Serve flooded buildings data to terria without styling
