### 2025-02-04 54acdab
Add colour to terria buildings using is_flooded
