### 2025-02-04 4e52823
Add is_flooded style to terria legend
