### 2025-02-04 0b843e8
Extract function for building flood status catalog item
