### 2025-02-18 28c7687
Include paragraph on state of the field (#349)
