+++ paper/paper.bib
@@ -172,3 +172,18 @@ eprint = {https://doi.org/10.1080/19475705.2018.1552630}
     author={Henderson, Roddy and Collins, Daniel and Doyle, Martin and Watson, Jeff},
     number={2018177CH},
 }
+
+@online{fastflood,
+    author = {Bastian van den Bout},
+    title = {FastFlood | FastFlood website, free super-fast flood modelling tool.},
+    year = {2024},
+    url = {https://fastflood.org/site/_site/},
+    urldate = {January 1 2025}
+}
+@online{flood-adapt,
+    author = {Kathryn Roscoe and Gundula Winter},
+    title = {FloodAdapt – an adaptation planning tool},
+    url = {https://www.deltares.nl/en/software-and-data/products/floodadapt},
+    urldate = {January 1 2025}
+}
+
