### 2025-02-18 #7433 @LukeParky
@cheginit I have completed the changes for the paper but I'm still awaiting feedback from last month's functional changes, after both reviewers encountered an issue

Issue links:
https://github.com/GeospatialResearch/Digital-Twins/issues/342
https://github.com/GeospatialResearch/Digital-Twins/issues/344
