### 2025-02-20 bedf27e
Fix bug causing issue with incorrect buildings layer
