### 2025-02-27 41f3ba3
Remove unused endpoints
