+++ floodresilience/blueprint.py
@@ -1,18 +1,10 @@
 """Endpoints and flask configuration for the Flood Resilience Digital Twin"""
 
-from http.client import OK, ACCEPTED, BAD_REQUEST, INTERNAL_SERVER_ERROR, NOT_FOUND
-import pathlib
-
-from celery import result, states
-from flask import Blueprint, Response, jsonify, make_response, send_file, request
+from flask import Blueprint
 from pywps import Service
-import requests
-from shapely import box
 
-from floodresilience import tasks
 from floodresilience.flood_model.flood_scenario_process_service import FloodScenarioProcessService
 from src.check_celery_alive import check_celery_alive
-from src.config import EnvVariable
 
 flood_resilience_blueprint = Blueprint('floodresilience', __name__)
 processes = [
@@ -36,296 +28,3 @@ def wps() -> Service:
         The PyWPS WebProcessing Service instance
     """
     return service
-
-
-@flood_resilience_blueprint.route('/tasks/<task_id>', methods=["GET"])
-def get_status(task_id: str) -> Response:
-    """
-    Retrieve status of a particular Celery backend task.
-    Supported methods: GET
-
-    Parameters
-    ----------
-    task_id : str
-        The id of the Celery task to retrieve status from
-
-    Returns
-    -------
-    Response
-        JSON response containing taskStatus
-    """
-    task_result = result.AsyncResult(task_id, app=tasks.app)
-    status = task_result.status
-    http_status = OK
-    if status == states.SUCCESS:
-        task_value = task_result.get()
-    elif status == states.FAILURE:
-        http_status = INTERNAL_SERVER_ERROR
-        is_debug_mode = EnvVariable.DEBUG_TRACEBACK
-        task_value = task_result.traceback if is_debug_mode else None
-    else:
-        task_value = None
-
-    return make_response(jsonify({
-        "taskId": task_result.id,
-        "taskStatus": status,
-        "taskValue": task_value
-    }), http_status)
-
-
-@flood_resilience_blueprint.route('/models/generate', methods=["POST"])
-@check_celery_alive
-def generate_model() -> Response:
-    """
-    Generate a flood model for a given area.
-    Supported methods: POST
-
-    Returns
-    -------
-    Response
-        ACCEPTED is the expected response. Response body contains Celery taskId
-    """
-    try:
-        bbox = request.get_json()["bbox"]
-        lat1 = float(bbox.get("lat1"))
-        lng1 = float(bbox.get("lng1"))
-        lat2 = float(bbox.get("lat2"))
-        lng2 = float(bbox.get("lng2"))
-        scenario_options = request.get_json()["scenarioOptions"]
-    except ValueError:
-        return make_response(
-            "JSON values for bbox: lat1, lng1, lat2, lng2 must be valid floats", BAD_REQUEST
-        )
-    if any(coord is None for coord in [lat1, lng1, lat2, lng2]):
-        return make_response("JSON body parameters bbox: {lat1, lng1, lat2, lng2} mandatory", BAD_REQUEST)
-    if not valid_coordinates(lat1, lng1) or not valid_coordinates(lat2, lng2):
-        return make_response("lat & lng must fall in the range -90 < lat <= 90, -180 < lng <= 180", BAD_REQUEST)
-    if (lat1, lng1) == (lat2, lng2):
-        return make_response("lat1, lng1 must not equal lat2, lng2", BAD_REQUEST)
-    is_valid_scenario, invalid_reason = tasks.validate_slr_parameters(scenario_options)
-    if not is_valid_scenario:
-        return make_response(invalid_reason, BAD_REQUEST)
-
[…]
