### 2025-03-02 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47, thank you for getting back to me. I understand this will be a very busy time for you so I think having a quick call might be the best way to make sure your time for this review is efficient.

I think that all you need to do is refresh your browser window, unless another issue is coming up further down the logs.

I'll check if a call like this is allowed with the editor, but in the meantime please let me know if refreshing the browser works.
