### 2025-03-02 #7433 @LukeParky
Hi @cheginit, I have been talking with @pritamd47 over on issue https://github.com/GeospatialResearch/Digital-Twins/issues/342 and he has encountered error messages and I would like to provide support for these over a video call to fix it more quickly.

Is it allowed for us to set up a call during the review, to resolve this quickly?
