### 2025-03-02 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi Luke,

I actually tried refreshing my window after seeing the error but got the
same error. Should I wait for a while before hitting refresh?

I’ll try again sometime today. If it doesn’t work, a call would be great to
see what could be going wrong and if it is something local to the system I
am using.

Pritam

On Sun, Mar 2, 2025 at 12:15 PM Luke Parkinson ***@***.***>
wrote:

> Hi @pritamd47 <https://github.com/pritamd47>, thank you for getting back
> to me. I understand this will be a very busy time for you so I think having
> a quick call might be the best way to make sure your time for this review
> is efficient.
>
> I think that all you need to do is refresh your browser window, unless
> another issue is coming up further down the logs.
>
> I'll check if a call like this is allowed with the editor, but in the
> meantime please let me know if refreshing the browser works.
>
