### 2025-03-03 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47, I have sent a Teams meeting invite to the email address on your github profile. Please let me know if it's suitable.
