### 2025-03-05 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi @LukeParky, thank you for getting on call to help me with setting up of the software. I am getting the same error of `PermissionError: [Error 13] Permission denied: '/stored_data/lidar'`. Here are the full logs -

```
celery_worker_digital_twin  | [2025-03-05 05:02:39,295: INFO/MainProcess] dataset table does not exist, initialising LiDAR dataset information.
celery_worker_digital_twin  | [2025-03-05 05:02:39,295: INFO/MainProcess] Start crawling datasets from OpenTopography.
celery_worker_digital_twin  | [2025-03-05 05:02:49,562: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'e75ee036-8df9-34bd-88c6-f130c1b3bf2a'} ticket:65385dc9-5b1d-465c-9d22-d69c61deaf15]
celery_worker_digital_twin  | [2025-03-05 05:03:04,061: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'dfe85883-c58c-33d7-8c43-f25169048daa'} ticket:17a8a07a-3d05-446d-80c4-c43f8f9b72b2]
celery_worker_digital_twin  | [2025-03-05 05:03:18,931: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'da409363-13e2-36d3-ba2d-0dcb9ec1597c'} ticket:bf49907d-be82-4855-b0a4-26b34b34e54e]
celery_worker_digital_twin  | [2025-03-05 05:03:34,527: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '8b25fb17-1e8e-344c-bcd5-115be11f6f3d'} ticket:ce1a6902-6306-46dd-9b25-f63b91155f88]
celery_worker_digital_twin  | [2025-03-05 05:03:49,423: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '1396396d-5a95-338f-b053-8ff8b4f06cdc'} ticket:23e55ab2-2fd7-4d34-af12-307dd6e6e3ec]
celery_worker_digital_twin  | [2025-03-05 05:04:04,704: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'ea180508-a82a-306a-a44e-0618cfe6cc48'} ticket:51274d6f-e002-44fb-a503-058976455d69]
celery_worker_digital_twin  | [2025-03-05 05:04:19,780: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '30c3acb5-834c-346d-96e5-a33230076a5a'} ticket:7a2d5a68-5f35-4203-89f1-dd52aa342249]
celery_worker_digital_twin  | [2025-03-05 05:04:34,740: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '69cd6299-40fe-3108-a8d8-04c39ae8e62e'} ticket:c6af476f-9a59-4c9a-8916-cdd91c85e0bc]
celery_worker_digital_twin  | [2025-03-05 05:04:49,600: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'd2dbf587-d214-36a3-a317-55ec20ffe720'} ticket:77c49fa3-f4e8-4dca-aee2-5366169d14a5]
celery_worker_digital_twin  | [2025-03-05 05:05:04,502: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': 'e66bf8b3-b0d7-3b7e-a90c-e80a39145b82'} ticket:9cd3369a-9cc0-41c2-9f0c-9161f7fe799a]
celery_worker_digital_twin  | [2025-03-05 05:05:19,824: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '9665ad6c-97c7-371b-9559-98eb4e4869ae'} ticket:6c7ac1a4-eb63-4b39-8d5f-372e60a5b039]
celery_worker_digital_twin  | [2025-03-05 05:05:34,842: DEBUG/MainProcess] pidbox received method ping() [reply_to:{'exchange': 'reply.celery.pidbox', 'routing_key': '34c414b9-b064-34eb-b698-3c83e133b1af'} ticket:223afcea-4526-4456-acc2-6b7559ef48cb]
celery_worker_digital_twin  | [2025-03-05 05:05:39,460: INFO/MainProcess] Finish crawling datasets from OpenTopography.
celery_worker_digital_twin  | [2025-03-05 05:05:39,461: WARNING/MainProcess] 2025-03-05 05:05:39 [newzealidar.datasets] INFO: Finish crawling datasets from OpenTopography.
celery_worker_digital_twin  | [2025-03-05 05:05:39,484: WARNING/MainProcess] --- Logging error ---
celery_worker_digital_twin  | [2025-03-05 05:05:39,485: WARNING/MainProcess] Traceback (most recent call last):
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 451, in trace_task
celery_worker_digital_twin  |     R = retval = fun(*args, **kwargs)
celery_worker_digital_twin  |                  ^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 734, in __protected_call__
celery_worker_digital_twin  |     return self.run(*args, **kwargs)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/app/src/tasks.py", line 142, in process_dem
celery_worker_digital_twin  |     process_hydro_dem.main(selected_polygon, **parameters)
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/app/src/flood_model/process_hydro_dem.py", line 121, in main
celery_worker_digital_twin  |     ensure_lidar_datasets_initialised()
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/app/src/flood_model/process_hydro_dem.py", line 49, in ensure_lidar_datasets_initialised
celery_worker_digital_twin  |     newzealidar.datasets.main()
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 416, in main
celery_worker_digital_twin  |     rename_file()
celery_worker_digital_twin  | [2025-03-05 05:05:39,486: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 384, in rename_file
celery_worker_digital_twin  |     data_dir.mkdir(parents=True, exist_ok=True)
celery_worker_digital_twin  | [2025-03-05 05:05:39,487: WARNING/MainProcess]   File "/venv/lib/python3.11/pathlib.py", line 1116, in mkdir
celery_worker_digital_twin  |     os.mkdir(self, mode)
celery_worker_digital_twin  | [2025-03-05 05:05:39,487: WARNING/MainProcess] PermissionError: [Errno 13] Permission denied: '/stored_data/lidar'
celery_worker_digital_twin  | [2025-03-05 05:05:39,487: WARNING/MainProcess]
celery_worker_digital_twin  | During handling of the above exception, another exception occurred:
celery_worker_digital_twin  | [2025-03-05 05:05:39,487: WARNING/MainProcess] Traceback (most recent call last):
celery_worker_digital_twin  | [2025-03-05 05:05:39,487: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 1110, in emit
celery_worker_digital_twin  |     msg = self.format(record)
celery_worker_digital_twin  |           ^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 953, in format
celery_worker_digital_twin  |     return fmt.format(record)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/utils/log.py", line 146, in format
celery_worker_digital_twin  |     msg = super().format(record)
celery_worker_digital_twin  |           ^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 695, in format
celery_worker_digital_twin  |     record.exc_text = self.formatException(record.exc_info)
celery_worker_digital_twin  |                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/utils/log.py", line 142, in formatException
celery_worker_digital_twin  |     r = super().formatException(ei)
celery_worker_digital_twin  |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 645, in formatException
celery_worker_digital_twin  |     traceback.print_exception(ei[0], ei[1], tb, None, sio)
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 124, in print_exception
celery_worker_digital_twin  |     te = TracebackException(type(value), value, tb, limit=limit, compact=True)
celery_worker_digital_twin  |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,488: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 688, in __init__
celery_worker_digital_twin  |     self.stack = StackSummary._extract_from_extended_frame_gen(
celery_worker_digital_twin  |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,489: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 416, in _extract_from_extended_frame_gen
celery_worker_digital_twin  |     for f, (lineno, end_lineno, colno, end_colno) in frame_gen:
celery_worker_digital_twin  | [2025-03-05 05:05:39,489: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 353, in _walk_tb_with_full_positions
celery_worker_digital_twin  |     positions = _get_code_position(tb.tb_frame.f_code, tb.tb_lasti)
celery_worker_digital_twin  |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,489: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 366, in _get_code_position
celery_worker_digital_twin  |     positions_gen = code.co_positions()
celery_worker_digital_twin  |                     ^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,489: WARNING/MainProcess] AttributeError: '_Code' object has no attribute 'co_positions'
celery_worker_digital_twin  | [2025-03-05 05:05:39,489: WARNING/MainProcess] Call stack:
celery_worker_digital_twin  | [2025-03-05 05:05:39,490: WARNING/MainProcess]   File "/venv/lib/python3.11/threading.py", line 995, in _bootstrap
celery_worker_digital_twin  |     self._bootstrap_inner()
celery_worker_digital_twin  | [2025-03-05 05:05:39,490: WARNING/MainProcess]   File "/venv/lib/python3.11/threading.py", line 1038, in _bootstrap_inner
celery_worker_digital_twin  |     self.run()
celery_worker_digital_twin  | [2025-03-05 05:05:39,490: WARNING/MainProcess]   File "/venv/lib/python3.11/threading.py", line 975, in run
celery_worker_digital_twin  |     self._target(*self._args, **self._kwargs)
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/concurrent/futures/thread.py", line 83, in _worker
celery_worker_digital_twin  |     work_item.run()
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/concurrent/futures/thread.py", line 58, in run
celery_worker_digital_twin  |     result = self.fn(*self.args, **self.kwargs)
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/concurrency/base.py", line 29, in apply_target
celery_worker_digital_twin  |     ret = target(*args, **kwargs)
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 649, in fast_trace_task
celery_worker_digital_twin  |     R, I, T, Rstr = tasks[task].__trace__(
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 468, in trace_task
celery_worker_digital_twin  |     I, R, state, retval = on_error(task_request, exc, uuid)
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 379, in on_error
celery_worker_digital_twin  |     R = I.handle_error_state(
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 178, in handle_error_state
celery_worker_digital_twin  |     return {
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 237, in handle_failure
celery_worker_digital_twin  |     self._log_error(task, req, einfo)
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 265, in _log_error
celery_worker_digital_twin  |     logger.log(policy.severity, policy.format.strip(), context,
celery_worker_digital_twin  | [2025-03-05 05:05:39,491: WARNING/MainProcess] Message: 'Task %(name)s[%(id)s] %(description)s: %(exc)s'
celery_worker_digital_twin  | Arguments: {'hostname': 'celery@9c5163f19e5b', 'id': '122b1b89-984e-453d-aede-d469a9336093', 'name': 'src.tasks.process_dem', 'exc': "PermissionError(13, 'Permission denied')", 'traceback': 'Traceback (most recent call last):\n  File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 451, in trace_task\n    R = retval = fun(*args, **kwargs)\n                 ^^^^^^^^^^^^^^^^^^^^\n  File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 734, in __protected_call__\n    return self.run(*args, **kwargs)\n           ^^^^^^^^^^^^^^^^^^^^^^^^^\n  File "/app/src/tasks.py", line 142, in process_dem\n    process_hydro_dem.main(selected_polygon, **parameters)\n  File "/app/src/flood_model/process_hydro_dem.py", line 121, in main\n    ensure_lidar_datasets_initialised()\n  File "/app/src/flood_model/process_hydro_dem.py", line 49, in ensure_lidar_datasets_initialised\n    newzealidar.datasets.main()\n  File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 416, in main\n    rename_file()\n  File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 384, in rename_file\n    data_dir.mkdir(parents=True, exist_ok=True)\n  File "/venv/lib/python3.11/pathlib.py", line 1116, in mkdir\n    os.mkdir(self, mode)\nPermissionError: [Errno 13] Permission denied: \'/stored_data/lidar\'\n', 'args': "['POLYGON ((172.7131337872427 -43.39377030256972, 172.7131337872427 -43.38426113151589, 172.7001528519404 -43.38426113151589, 172.7001528519404 -43.39377030256972, 172.7131337872427 -43.39377030256972))']", 'kwargs': '{}', 'description': 'raised unexpected', 'internal': False}
celery_worker_digital_twin  | [2025-03-05 05:05:39,492: WARNING/MainProcess] --- Logging error ---
celery_worker_digital_twin  | [2025-03-05 05:05:39,493: WARNING/MainProcess] Traceback (most recent call last):
celery_worker_digital_twin  | [2025-03-05 05:05:39,494: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 451, in trace_task
celery_worker_digital_twin  |     R = retval = fun(*args, **kwargs)
celery_worker_digital_twin  |                  ^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,494: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/celery/app/trace.py", line 734, in __protected_call__
celery_worker_digital_twin  |     return self.run(*args, **kwargs)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,494: WARNING/MainProcess]   File "/app/src/tasks.py", line 142, in process_dem
celery_worker_digital_twin  |     process_hydro_dem.main(selected_polygon, **parameters)
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/app/src/flood_model/process_hydro_dem.py", line 121, in main
celery_worker_digital_twin  |     ensure_lidar_datasets_initialised()
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/app/src/flood_model/process_hydro_dem.py", line 49, in ensure_lidar_datasets_initialised
celery_worker_digital_twin  |     newzealidar.datasets.main()
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 416, in main
celery_worker_digital_twin  |     rename_file()
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/site-packages/newzealidar/datasets.py", line 384, in rename_file
celery_worker_digital_twin  |     data_dir.mkdir(parents=True, exist_ok=True)
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/pathlib.py", line 1116, in mkdir
celery_worker_digital_twin  |     os.mkdir(self, mode)
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess] PermissionError: [Errno 13] Permission denied: '/stored_data/lidar'
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]
celery_worker_digital_twin  | During handling of the above exception, another exception occurred:
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess] Traceback (most recent call last):
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 1110, in emit
celery_worker_digital_twin  |     msg = self.format(record)
celery_worker_digital_twin  |           ^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 953, in format
celery_worker_digital_twin  |     return fmt.format(record)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,495: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 695, in format
celery_worker_digital_twin  |     record.exc_text = self.formatException(record.exc_info)
celery_worker_digital_twin  |                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/logging/__init__.py", line 645, in formatException
celery_worker_digital_twin  |     traceback.print_exception(ei[0], ei[1], tb, None, sio)
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 124, in print_exception
celery_worker_digital_twin  |     te = TracebackException(type(value), value, tb, limit=limit, compact=True)
celery_worker_digital_twin  |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 688, in __init__
celery_worker_digital_twin  |     self.stack = StackSummary._extract_from_extended_frame_gen(
celery_worker_digital_twin  |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 416, in _extract_from_extended_frame_gen
celery_worker_digital_twin  |     for f, (lineno, end_lineno, colno, end_colno) in frame_gen:
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 353, in _walk_tb_with_full_positions
celery_worker_digital_twin  |     positions = _get_code_position(tb.tb_frame.f_code, tb.tb_lasti)
celery_worker_digital_twin  |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess]   File "/venv/lib/python3.11/traceback.py", line 366, in _get_code_position
celery_worker_digital_twin  |     positions_gen = code.co_positions()
celery_worker_digital_twin  |                     ^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess] AttributeError: '_Code' object has no attribute 'co_positions'
celery_worker_digital_twin  | [2025-03-05 05:05:39,496: WARNING/MainProcess] Call stack:
[…]
