### 2025-03-05 GeospatialResearch/Digital-Twins#342 @LukeParky
Hi @pritamd47, to confirm this problem, could you please let me know the OS you were running this in? I think from our call yesterday it was running on WSL so you can get the OS through `cat /etc/os-release`.
