### 2025-03-05 GeospatialResearch/Digital-Twins#342 @LukeParky
Okay, I can confirm this happens in other OSs, its much more closely related to #344 than the issue that you originally opened. I believe this issue (Increase explanations for POSTGRES_PASSWORD .env) is fixed, but I will reopen #344.
