### 2025-03-05 GeospatialResearch/Digital-Twins#344 @LukeParky
As mentioned in #342, this issue occurs in Ubuntu-22.04 and possibly other operations systems. I am reopening the issue to provide a fix. I believe this is fixed in downstream forks so will try to isolate the fix for this release.
