### 2025-03-07 9e86aa1
Use docker-compose.yml format from further upstream
