### 2025-03-07 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi @LukeParky, the OS I was running was Ubuntu. I see that you have reopened the issue #344. I will follow that thread for updates on the fix. Let me know when you're able to push fix, and I will rerun the application.
