### 2025-03-11 GeospatialResearch/Digital-Twins#342 @LukeParky
> Hi [@LukeParky](https://github.com/LukeParky), the OS I was running was Ubuntu. I see that you have reopened the issue [#344](https://github.com/GeospatialResearch/Digital-Twins/issues/344). I will follow that thread for updates on the fix. Let me know when you're able to push fix, and I will rerun the application.

Hi @pritamd47, could you let me know the exact version of Ubuntu you were running? I want to complete some more testing on that version specifically
