### 2025-03-12 a894e55
344 permission denied docker fix (#350)
* Bump docker image versions, to ensure that cached instances are updated on other devices

* Use docker-compose.yml format from further upstream
