### 2025-03-12 GeospatialResearch/Digital-Twins#344 @LukeParky
@changliao1025 please let me know if `git pull && docker compose pull && docker compose down -v && docker compose up -d && docker compose logs -f backend celery_worker` resolves this issue.
