### 2025-03-18 a6419bb
Add 3D height to buildings using simplestyle-spec and cesium primitives (#3)
* Add 3D height to buildings using simplestyle-spec and cesium primitives

* Use dedicated file for HTML/Mustache infobox
