+++ floodresilience/flood_model/flood_scenario_process_service.py
@@ -118,9 +118,9 @@ def building_flood_status_catalog(scenario_id: int) -> dict:
     dataset_name = "Building Flood Status"
     gs_building_workspace = f"{EnvVar.POSTGRES_DB}-buildings"
     gs_building_url = f"{EnvVar.GEOSERVER_HOST}:{EnvVar.GEOSERVER_PORT}/geoserver/{gs_building_workspace}/ows"
-
-    flooded_color = "darkred"
-    non_flooded_color = "darkgreen"
+    # Open and read HTML/mustache template file for infobox
+    with open("./floodresilience/flood_model/templates/flooded_building_infobox.mustache", encoding="utf-8") as file:
+        flooded_building_infobox_template = file.read()
     return {
         "type": "wfs",
         "name": dataset_name,
@@ -130,43 +130,24 @@ def building_flood_status_catalog(scenario_id: int) -> dict:
             "viewparams": f"scenario:{scenario_id}",
         },
         "maxFeatures": 300000,
-        "styles": [{
-            "id": "is_flooded",
-            "title": dataset_name,
-            "color": {
-                "mapType": "enum",
-                "colorColumn": "is_flooded_int",
-                "legend": {
-                    "title": dataset_name,
-                    "items": [
-                        {
-                            "title": "Non-Flooded",
-                            "color": non_flooded_color
-                        },
-                        {
-                            "title": "Flooded",
-                            "color": flooded_color
-                        }
-                    ]
+        "heightProperty": "extruded_height",
+        "featureInfoTemplate": {
+            "name": "Building Flood Status - {{flood_model_id}} - {{building_outline_id}}",
+            "template": flooded_building_infobox_template
+        },
+        "legends": [{
+            "title": "Building Flood Status",
+            "items": [
+                {
+                    "title": "Non-Flooded",
+                    "color": "darkgreen"
                 },
-                "enumColors": [
-                    {
-                        "value": "0",
-                        "color": non_flooded_color
-                    },
-                    {
-                        "value": "1",
-                        "color": flooded_color
-                    }
-                ]
-            },
-            "outline": {
-                "null": {
-                    "width": 0
+                {
+                    "title": "Flooded",
+                    "color": "darkred"
                 }
-            }
-        }],
-        "activeStyle": "is_flooded"
+            ]
+        }]
     }
 
 
