### 2025-03-20 96ca47e
Precache LiDAR on celery_worker startup (#8)
