+++ floodresilience/blueprint.py
@@ -1,8 +1,10 @@
 """Endpoints and flask configuration for the Flood Resilience Digital Twin"""
 
-from flask import Blueprint
+from http.client import ACCEPTED
+from flask import Blueprint, jsonify, make_response, Response
 from pywps import Service
 
+from floodresilience import tasks
 from floodresilience.flood_model.flood_scenario_process_service import FloodScenarioProcessService
 from src.check_celery_alive import check_celery_alive
 
@@ -28,3 +30,26 @@ def wps() -> Service:
         The PyWPS WebProcessing Service instance
     """
     return service
+
+
+@flood_resilience_blueprint.route('/datasets/update', methods=["POST"])
+@check_celery_alive
+def refresh_lidar_data_sources() -> Response:
+    """
+    Update LiDAR data sources to the most recent.
+    Web-scrape OpenTopography metadata to update the datasets table containing links to LiDAR data sources.
+    Takes a long time to run but needs to be run periodically so that the datasets are up to date.
+    Supported methods: POST
+
+    Returns
+    -------
+    Response
+        ACCEPTED is the expected response. Response body contains Celery taskId
+    """
+    # Start task to refresh lidar datasets
+    task = tasks.refresh_lidar_datasets.delay()
+    # Return HTTP Response with task id, so it can be monitored with get_status(taskId)
+    return make_response(
+        jsonify({"taskId": task.id}),
+        ACCEPTED
+    )
