### 2025-03-24 GeospatialResearch/Digital-Twins#344 @changliao1025
It seems this issue still remains.
I am using an iMac with an Intel processor.
