### 2025-03-27 #7433 @LukeParky
@cheginit As far as I am aware the only remaining issues is #344. Both reviewers had encountered the issue, which prevents the application from running.

I pushed a fix 2 weeks ago and have not been able to replicate the issue since.
It seems @changliao1025 is still encountering it so I am investigating why that could be.
