### 2025-03-27 GeospatialResearch/Digital-Twins#344 @LukeParky
Hi @changliao1025 it is unfortunate you are still encountering the issue. I am not yet sure how it can be happening.

I have not been able to replicate it since pushing my fix.
I am attempting one more way to replicate this.

Could you please provide me with the contents of your `docker compose images` output?

E.g.
```
CONTAINER                     REPOSITORY                                 TAG                 IMAGE ID            SIZE
backend_digital_twin          lparkinson/flood-resilience-dt             1.4                 ae586e           2.93GB
celery_worker_digital_twin    lparkinson/flood-resilience-dt             1.4                 ae586e           2.93GB
db_postgres_digital_twin      postgis/postgis                            16-3.4              06287eb8e12c        609MB
geoserver_digital_twin        lparkinson/geoserver-flood-resilience-dt   1.4                 efc8ef37f170        1.11GB
message_broker_digital_twin   redis                                      7                   ad4b31aa2de6        117MB
www_digital_twin              lparkinson/www-flood-resilience-dt         1.4                 96b584736ee6        239MB
```
