### 2025-04-08 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi @LukeParky. Please see the latest logs. The celery worker is getting killed, and I am facing the same issue as before "Could not reach backend. Please refresh the page to try again."

![Image](https://github.com/user-attachments/assets/1f13bfe5-0df6-4bf1-abb2-8055a8134e93)
