### 2025-04-08 GeospatialResearch/Digital-Twins#342 @LukeParky
I have seen this happen when the computer runs out of memory, do you think this could be the case?
