### 2025-04-09 GeospatialResearch/Digital-Twins#342 @pritamd47
Yes, that was likely the case. I cleared the memory and ran again after which it ran smoothly. However, I am still getting the `permission_denied` error #344:

![Image](https://github.com/user-attachments/assets/fec17726-6a76-403c-b32c-05e65e58fa7e)

I updated the repository using `git pull` before running any of the docker commands.
