### 2025-04-09 GeospatialResearch/Digital-Twins#342 @pritamd47
Hi Luke, yes these are the commands that I used for starting up the server.

On Wed, Apr 9, 2025 at 1:30 PM Luke Parkinson ***@***.***>
wrote:

> I updated the repository using git pull before running any of the docker
> commands.
>
> Can you confirm that the docker commands you ran were as stated in #344
> <https://github.com/GeospatialResearch/Digital-Twins/issues/344>, I am
> not sure how to tag you in that issue.
>
> You need new volumes because of the bug so you need to run:
> git pull && docker compose pull && docker compose down -v && docker
> compose up -d && docker compose logs -f backend celery_worker
>
> Can you please confirm this? Because if the error still occurs after
> running this then there is something else major causing issues that I
> cannot replicate.
>
