### 2025-04-09 GeospatialResearch/Digital-Twins#344 @LukeParky
In addition to the `docker compose images` output could I please also have the output of `docker volume inspect digital-twins_stored_data`?
