### 2025-04-09 GeospatialResearch/Digital-Twins#342 @LukeParky
@pritamd47 To help me solve this problem could you please give me the output of the following two commands:

`docker compose images`

`docker volume inspect digital-twins_stored_data`
Thank you
