### 2025-04-09 #7433 @LukeParky
@pritamd47 reported that he is still encountering a major error GeospatialResearch/Digital-Twins#344. I provided a fix in the code, and instructions on how to reset volumes to fix the issue.

@pritamd47 reports that it is still occurring, and I am waiting on additional info from @changliao1025.

I have not been able to replicate the issue after following the instructions I posted, so I may need to cancel this submission.

In order to fix this I would still like to receive some log information, which I have requested in that issue, and also in the original issue that @pritamd47 posted,  GeospatialResearch/Digital-Twins#342
