### 2025-04-17 GeospatialResearch/Digital-Twins#344 @changliao1025
This is the output from `docker compose images`:

This is the output from `docker volume inspect digital-twins_stored_data`:
