### 2025-04-17 GeospatialResearch/Digital-Twins#344 @changliao1025
Additional log info:
[2025-04-17 22:14:47,331: CRITICAL/MainProcess]
celery_worker_digital_twin  | Traceback (most recent call last):
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/twisted/internet/defer.py", line 2017, in _inlineCallbacks
celery_worker_digital_twin  |     result = context.run(gen.send, result)
celery_worker_digital_twin  |              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/crawler.py", line 158, in crawl
celery_worker_digital_twin  |     self.engine = self._create_engine()
celery_worker_digital_twin  |                   ^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/crawler.py", line 172, in _create_engine
celery_worker_digital_twin  |     return ExecutionEngine(self, lambda _: self.stop())
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/core/engine.py", line 100, in __init__
celery_worker_digital_twin  |     self.scraper = Scraper(crawler)
celery_worker_digital_twin  |                    ^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/core/scraper.py", line 109, in __init__
celery_worker_digital_twin  |     self.itemproc: ItemPipelineManager = itemproc_cls.from_crawler(crawler)
celery_worker_digital_twin  |                                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/middleware.py", line 90, in from_crawler
celery_worker_digital_twin  |     return cls.from_settings(crawler.settings, crawler)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/middleware.py", line 67, in from_settings
celery_worker_digital_twin  |     mw = create_instance(mwcls, settings, crawler)
celery_worker_digital_twin  |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/utils/misc.py", line 188, in create_instance
celery_worker_digital_twin  |     instance = objcls.from_crawler(crawler, *args, **kwargs)
celery_worker_digital_twin  |                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/media.py", line 78, in from_crawler
celery_worker_digital_twin  |     pipe = cls.from_settings(crawler.settings)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/files.py", line 391, in from_settings
celery_worker_digital_twin  |     return cls(store_uri, settings=settings)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/files.py", line 351, in __init__
celery_worker_digital_twin  |     self.store = self._get_store(store_uri)
celery_worker_digital_twin  |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/files.py", line 399, in _get_store
celery_worker_digital_twin  |     return store_cls(uri)
celery_worker_digital_twin  |            ^^^^^^^^^^^^^^
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/files.py", line 55, in __init__
celery_worker_digital_twin  |     self._mkdir(Path(self.basedir))
celery_worker_digital_twin  |   File "/venv/lib/python3.11/site-packages/scrapy/pipelines/files.py", line 85, in _mkdir
celery_worker_digital_twin  |     dirname.mkdir(parents=True)
celery_worker_digital_twin  |   File "/venv/lib/python3.11/pathlib.py", line 1116, in mkdir
celery_worker_digital_twin  |     os.mkdir(self, mode)
celery_worker_digital_twin  | PermissionError: [Errno 13] Permission denied: '/stored_data/lidar'
