### 2025-04-18 GeospatialResearch/Digital-Twins#342 @pritamd47
Output of `docker compose images` ![Image](https://github.com/user-attachments/assets/8db15552-927e-430e-a2a3-c9c2439a762e)

Output of `docker volume inspect digital-twins_stored_data`
![Image](https://github.com/user-attachments/assets/810a4715-c636-4390-9707-a48967b0f916)
