### 2025-04-18 #7433 @cheginit
@pritamd47 and @changliao1025 thank you both for the update, appreciate your effort in trying to provide more information on the issues.

@LukeParky, please check the new information provided by the reviewers to address the issues.
