+++ wrfhydro/land_surface_model/create_land_surface_model.py
@@ -1,0 +1,39 @@
+# # -*- coding: utf-8 -*-
+# Copyright © 2021-2025 Geospatial Research Institute Toi Hangarau
+# LICENSE: https://github.com/GeospatialResearch/Digital-Twins/blob/master/LICENSE
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Create a land surface model for use within WRF-Hydro modelling."""
+
+import geopandas as gpd
+
+from src.digitaltwin.utils import LogLevel
+
+
+def main(
+    selected_polygon_gdf: gpd.GeoDataFrame,
+    log_level: LogLevel = LogLevel.DEBUG
+) -> None:
+    """
+    Create a land surface model for use within WRF-Hydro modelling.
+
+    Parameters
+    ----------
+    selected_polygon_gdf : gpd.GeoDataFrame
+        A GeoDataFrame representing the selected polygon, i.e., the catchment area.
+    log_level : LogLevel = LogLevel.DEBUG
+        The log level to set for the root logger. Defaults to LogLevel.DEBUG.
+    """
+    pass
