### 2025-05-16 #7433 @LukeParky
@cheginit I've focused my efforts to get a new release out. I've just pushed the release v2.0.0, and cannot find any evidence of the bug that both @pritamd47 and @changliao1025 encountered.

I was not able to test on MacOSX but tested extensively on Windows, WSL, and Ubuntu.

@pritamd47, @changliao1025 thank you for your patience. The code is ready to test again. I will add instructions to https://github.com/GeospatialResearch/Digital-Twins/issues/344.
