### 2025-07-07 #7433 @LukeParky
It seems like the remaining issue may be MacOS-specific but we do not have easy access to a Mac device.

We are trying to find a solution for Mac, but the alternative may be to explicitly not support MacOS.
