### 2025-07-07 #7433 @LukeParky
We have made the decision that we can not support MacOS due to lack of access to devices.
This may impact @changliao1025 's review but we are hoping reviews can be completed now.

@pritamd47 are you able to provide another update?
