### 2025-07-09 #7433 @pritamd47
Hi @LukeParky , I updated the issue at https://github.com/GeospatialResearch/Digital-Twins/issues/344#issuecomment-3050865439.

I am facing another issue now. If this issue is related to my personal computer (that it keeps crashing while building the docker image), I don't have another computer to try out the fix unfortunately. I had extra machines I could attempt to try out the fix earlier but I have moved countries after graduating, and no longer have access to those machines/people. @cheginit Kindly advise what would be the best way to move forward, considering also that other issues have remained in the software (discussed at length at https://github.com/GeospatialResearch/Digital-Twins/issues/344#issuecomment-3050865439 ). Would it make more sense to do a fresh submission after the issues encountered in a fresh install of the software are thoroughly ironed out?
