### 2025-07-22 #7433 @LukeParky
> [@LukeParky](https://github.com/LukeParky) It sounds like the installation issues have not been resolved. Considering that, it is a critical step in testing and reviewing the functionalities of your submission, would you like to withdraw?

@cheginit While we believe the major issues have been resolved we agree that further testing is needed to ensure the stability.
We are going to be conducting testing with a wider range of devices, including using students who have less experience setting up this form of software, in order to improve our process.

Unfortunately we won't be able to support Mac OSX which has been a recurring barrier in this review.

We would like to withdraw and resubmit once this is all completed.
